### 2023-06-09 #114 @jagerber48
Hello @NickleDave, thanks for your response!

1. Yes, the Pangeo affiliation was a misunderstanding on my part. I thought pyopensci had a requirement to be conformant with Pangeo in addition to pyopensci, so I was indicating willingness to conform there as well. I've unchecked that box.

2. Good question. I will think about this but here are my off-the-cuff thoughts.
- My main use case for this sort of formatting is when I am printing data analysis results to a terminal (either in some console program) or in a jupyter notebook. I'm trying to view data analysis results part way through a workflow or at the end of the workflow and I would like to see them, e.g. in engineering notation rounded to 3 significant figures. For this case the main "consumers" of `sciform` outputs are python `print` or `logger` functions. Sometimes I make a table using the tabulate package, fill it with formatted strings, and print that to the terminal.
- The next important use case would be integrating these formatted strings into plots in a variety of ways. One very nice example would be to plot tick labels using prefix notation. e.g., instead of having "1e3" out at the edge of the axis, you would have tick labels like 1 k, 2 k, 10 k, 50 k, etc. of course including strings like this on annotations could be useful as well.
- The uncertainties package provides the ability to translate formatted value +/- uncertainty strings into latex code, but I need to understand what the use cases are there. I'm not advanced enough in latex to understand how someone would integrate python data analysis directly into latex source code for a document. I would just copy the number into latex manually and round/format them as needed, but maybe there's a more sophisticated workflow I'm missing.
- There might be use cases for writing formatted data into e.g. csv or other types of files for storing data. But I would suggest that if you are *storing* numbers you should store them appropriately as ints or floats, not strings.

I can include the tabulate and matplotlib use cases in the documentation. I think those would be illustrative use cases people could look at.
