+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.0"
+__version__ = "0.12.1"
 
 from sciform.formatter import Formatter
 from sciform.sfloat import sfloat
