+++ src/sciform/unc_format.py
@@ -1,4 +1,6 @@
 import re
+from math import isfinite
+from warnings import warn
 
 from sciform.modes import RoundMode, SignMode, FormatMode
 from sciform.format_utils import (get_round_digit, get_top_and_bottom_digit,
@@ -8,15 +10,23 @@ from sciform.format_options import FormatOptions
 from sciform.prefix import replace_prefix
 
 
-def format_val_unc(val: float, unc: float, fmt: str):
+def format_val_unc(val: float, unc: float, options: FormatOptions,
+                   match_widths: bool = False,
+                   paren_unc: bool = False):
+    """
+    Convert two floats, the value and the uncertainty into a
+
+    :param val:
+    :param unc:
+    :param options:
+    :param match_widths:
+    :return:
+    """
     # TODO: Handle non-finite floats
-    options = FormatOptions.from_format_spec_str(fmt)
-    pattern = FormatOptions.pattern
-    match = re.match(pattern, fmt)
-    if match.group('top_dig_place'):
-        match_widths = True
-    else:
-        match_widths = False
+    if options.round_mode is RoundMode.PREC:
+        warn('Precision round mode not available for value/uncertainty '
+             'formatting. Rounding is always applied as significant figures'
+             'for the uncertainty.')
 
     unc = abs(unc)
 
@@ -24,77 +34,118 @@ def format_val_unc(val: float, unc: float, fmt: str):
         val *= 100
         unc *= 100
 
-    unc_top_digit, unc_bottom_digit = get_top_and_bottom_digit(unc)
-    round_digit = get_round_digit(unc_top_digit, unc_bottom_digit,
+    # Find the digit place to round to
+    if isfinite(unc) and unc != 0:
+        round_driver = unc
+    else:
+        round_driver = val
+
+    top_digit_1, bottom_digit_1 = get_top_and_bottom_digit(round_driver)
+    round_digit = get_round_digit(top_digit_1, bottom_digit_1,
                                   options.precision, RoundMode.SIG_FIG)
 
+    # Perform rounding
     unc_rounded = round(unc, -round_digit)
     val_rounded = round(val, -round_digit)
+    round_driver = round(round_driver, -round_digit)
 
     '''
     Re-round the rounded values in case the first rounding changed the most
     significant digit place.
     '''
-    unc_top_digit, unc_bottom_digit = get_top_and_bottom_digit(unc_rounded)
-    round_digit = get_round_digit(unc_top_digit, unc_bottom_digit,
+    top_digit_2, bottom_digit_2 = get_top_and_bottom_digit(round_driver)
+    round_digit = get_round_digit(top_digit_2, bottom_digit_2,
                                   options.precision, RoundMode.SIG_FIG)
 
     unc_rounded = round(unc_rounded, -round_digit)
     val_rounded = round(val_rounded, -round_digit)
 
-    val_mantissa, exp, base = get_mantissa_exp_base(
-        val_rounded,
-        format_mode=options.format_mode,
-        exp=options.exp)
 
-    unc_mantissa, _, _ = get_mantissa_exp_base(
-        unc_rounded,
-        format_mode=options.format_mode,
-        exp=exp)
-    unc_top_digit, unc_bottom_digit = get_top_and_bottom_digit(unc_mantissa)
-    round_digit = get_round_digit(unc_top_digit, unc_bottom_digit,
-                                  options.precision, options.round_mode)
+    format_mode = options.format_mode
 
-    prec = -round_digit
+    '''
+    Get a corresponding format mode which can have the exponent set
+    explicitly.
+    '''
+    if format_mode is FormatMode.PERCENT:
+        free_exp_format_mode = FormatMode.FIXEDPOINT
[…]

[…]
