### 2023-06-10 da97959
remove some now redundant tests
