+++ tests/test_val_unc.py
@@ -168,41 +168,6 @@ class TestFormatting(unittest.TestCase):
 
         self.do_unc_val_test_case_dict(cases_dict)
 
-    def test_various(self):
-        cases_dict = {
-            (123.456, 0.789):
-                {
-                    '': '123.456 +/- 0.789',
-                    'f': '123.456 +/- 0.789',
-                    'e': '(1.23456 +/- 0.00789)e+02',
-                    'r': '(123.456 +/- 0.789)e+00',
-                    '#r': '(0.123456 +/- 0.000789)e+03',
-                    'R': '(123.456 +/- 0.789)E+00',
-                    'S': '123.456(789)',
-                    'fS': '123.456(789)',
-                    'eS': '(1.23456(789))e+02',
-                    'rS': '(123.456(789))e+00',
-                    'RS': '(123.456(789))E+00',
-                    '!3': '123.456 +/- 0.789',
-                    '!3f': '123.456 +/- 0.789',
-                    '!3e': '(1.23456 +/- 0.00789)e+02',
-                    '!3r': '(123.456 +/- 0.789)e+00',
-                    '!3R': '(123.456 +/- 0.789)E+00',
-                    '+': '+123.456 +/- 0.789',
-                    '+f': '+123.456 +/- 0.789',
-                    '+e': '(+1.23456 +/- 0.00789)e+02',
-                    '+r': '(+123.456 +/- 0.789)e+00',
-                    '+R': '(+123.456 +/- 0.789)E+00',
-                    ' ': ' 123.456 +/- 0.789',
-                    ' f': ' 123.456 +/- 0.789',
-                    ' e': '( 1.23456 +/- 0.00789)e+02',
-                    ' r': '( 123.456 +/- 0.789)e+00',
-                    ' R': '( 123.456 +/- 0.789)E+00',
-                }
-        }
-
-        self.do_unc_val_test_case_dict(cases_dict)
-
     def test_nan_inf(self):
         cases_dict = {
             (12.34, NAN): {
