+++ src/sciform/modes.py
@@ -11,6 +11,10 @@ class AutoPrec:
     pass
 
 
+class AutoValUncNanIncludeExp:
+    pass
+
+
 def option_warn_str(value, options: type(Literal)):
     return f'Flag \'{value}\' not in {get_args(options)}.'
 
