### 2023-06-10 a48d412
move match_widths and paren/bracket_unc into options class
