+++ tests/test_val_unc.py
@@ -199,9 +199,30 @@ class TestFormatting(unittest.TestCase):
     def test_rounding(self):
         cases_dict = {
             (0.0999, 0.0999): {
-                '!1e': '(1 +/- 1)e-01'},
+                '!1e': '(1 +/- 1)e-01'
+            },
             (0.0999, 0.999): {
-                '!1e': '(0 +/- 1)e+00'}
+                '!1e': '(0 +/- 1)e+00'
+            }
+        }
+        self.do_unc_val_test_case_dict(cases_dict)
+
+    def test_match_wdiths(self):
+        cases_dict = {
+            (123.456, 0.789): {
+                '0=0': '123.456 +/- 000.789',
+                '0= 0': ' 123.456 +/- 000.789',
+                ' 0': ' 123.456 +/-   0.789',
+                '+0': '+123.456 +/-   0.789',
+                '0': '123.456 +/-   0.789'
+            },
+            (0.789, 123.456): {
+                '0=0': '000.789 +/- 123.456',
+                '0= 0': ' 000.789 +/- 123.456',
+                ' 0': '   0.789 +/- 123.456',
+                '+0': '+  0.789 +/- 123.456',
+                '0': '  0.789 +/- 123.456'
+            }
         }
         self.do_unc_val_test_case_dict(cases_dict)
 
