### 2023-06-10 4058b7f
logic for val_unc_nan_include_exp and options copied all over
