+++ src/sciform/format_options.py
@@ -27,12 +27,12 @@ class FormatOptions:
     use_prefix: bool
     extra_si_prefixes: dict[int, str]
     extra_iec_prefixes: dict[int, str]
-    val_unc_match_widths: bool
-    nan_include_exp: bool
-    val_unc_nan_include_exp: Union[bool, type(AutoValUncNanIncludeExp)]
     bracket_unc: bool
+    val_unc_match_widths: bool
     bracket_unc_remove_dec_symb: bool
     unc_pm_whitespace: bool
+    nan_include_exp: bool
+    val_unc_nan_include_exp: Union[bool, type(AutoValUncNanIncludeExp)]
 
     def __post_init__(self):
         if self.round_mode is RoundMode.SIG_FIG:
@@ -98,12 +98,12 @@ class FormatOptions:
             extra_iec_prefixes: dict[int, str] = None,
             add_c_prefix: bool = False,
             add_small_si_prefixes: bool = False,
+            bracket_unc: bool = None,
             val_unc_match_widths: bool = None,
+            bracket_unc_remove_dec_symb: bool = None,
+            unc_pm_whitespace: bool = None,
             nan_include_exp: bool = None,
             val_unc_nan_include_exp: Union[bool, type(AutoValUncNanIncludeExp)] = None,
-            bracket_unc: bool = None,
-            bracket_unc_remove_dec_symb: bool = None,
-            unc_pm_whitespace: bool = None
     ):
         if defaults is None:
             defaults = DEFAULT_GLOBAL_OPTIONS
@@ -144,15 +144,15 @@ class FormatOptions:
             extra_si_prefixes=extra_si_prefixes,
             extra_iec_prefixes=(copy(extra_iec_prefixes)
                                 or copy(defaults.extra_iec_prefixes)),
+            bracket_unc=bracket_unc or defaults.bracket_unc,
             val_unc_match_widths=(val_unc_match_widths or
                                   defaults.val_unc_match_widths),
+            bracket_unc_remove_dec_symb=(bracket_unc_remove_dec_symb or
+                                         defaults.bracket_unc_remove_dec_symb),
+            unc_pm_whitespace=unc_pm_whitespace or defaults.unc_pm_whitespace,
             nan_include_exp=nan_include_exp or defaults.nan_include_exp,
             val_unc_nan_include_exp=(val_unc_nan_include_exp or
                                      defaults.val_unc_nan_include_exp),
-            bracket_unc=bracket_unc or defaults.bracket_unc,
-            bracket_unc_remove_dec_symb=(bracket_unc_remove_dec_symb or
-                                         defaults.bracket_unc_remove_dec_symb),
-            unc_pm_whitespace=unc_pm_whitespace or defaults.unc_pm_whitespace
         )
 
     pattern = re.compile(r'''^
@@ -284,8 +284,8 @@ class FormatOptions:
             capital_exp_char=capital_exp_char,
             exp=exp,
             use_prefix=use_prefix,
-            val_unc_match_widths=val_unc_match_widths,
-            bracket_unc=bracket_unc
+            bracket_unc=bracket_unc,
+            val_unc_match_widths=val_unc_match_widths
         )
 
 
@@ -304,12 +304,12 @@ DEFAULT_PKG_OPTIONS = FormatOptions(
     use_prefix=False,
     extra_si_prefixes=dict(),
     extra_iec_prefixes=dict(),
-    val_unc_match_widths=False,
-    nan_include_exp=False,
-    val_unc_nan_include_exp=AutoValUncNanIncludeExp,
     bracket_unc=False,
+    val_unc_match_widths=False,
     bracket_unc_remove_dec_symb=False,
-    unc_pm_whitespace=True
+    unc_pm_whitespace=True,
[…]
