### 2023-06-11 65d601d
bug on exp_str for non-finite val + tests
