### 2023-06-14 #114 @jagerber48
@lwasser

About the Pangeo option from my perspective:

> We partner with communities to support peer review with an additional layer of
checks that satisfy community requirements. If your package fits into an
existing community please check below:

>
> [Pangeo](https://www.pangeo.io/)
> My package adheres to the [Pangeo standards listed in the pyOpenSci peer review guidebook](https://www.pyopensci.org/software-peer-review/partners/pangeo)

"If your package fits into an existing community please check below" was a challenging sentence for me. I have no idea what Pangeo is, and didn't really want to research it to have to make this pre-submission. I still don't really know what Pangeo is or the details of this "community partnership" so it's hard for me to propose amended verbiage here for what would make it more clear, but for me, just more words about what the partnership means (are there benefits? are there costs?) would be helpful. An "(optional)" flag would be an easy way to help I guess?

> Also i'm wondering then if this tool would really be a support tool for reproducible reports (which is important to our open science goals)? If it's really about printing and output. Does that type of application (reproducible reports/ jupyter notebook output, etc). resonate with your goals for the tool?

"would really be a support tool for reproducible reports". What are "reproducible reports"? The tool takes python floats or float pairs and converts them to formatted (hopefully human readable) strings. There are many ways these strings could be used, it sounds like "reproducible reports" is definitely a use case that this tool could support. You mention Jupyter notebook output, that's definitely something I use it for, so I would say this does resonate with my goals for the tool.
