### 2023-06-16 d31ac21
clean up FormatOptions.make defaults extraction
