### 2023-06-16 fd58277
formatter use bracket_unc_remove_dec_symb
