### 2023-06-16 3c8e865
use unc_pm_whitespace and bracket_unc_remove_dec_symb
