### 2023-06-16 3b0641d
Unify float and val/unc formatting
