### 2023-06-16 6e3eef8
update print global defaults doctest
