### 2023-06-16 1d8c16d
bracket_unc_remove_seps rename and functionality
