+++ src/sciform/formatter.py
@@ -99,6 +99,7 @@ class Formatter:
             format_mode: FormatMode = None,
             capital_exp_char: bool = None,
             exp: Union[int, type(AutoExp)] = None,
+            nan_include_exp: bool = None,
             use_prefix: bool = None,
             extra_si_prefixes: dict[int, str] = None,
             extra_iec_prefixes: dict[int, str] = None,
@@ -108,7 +109,6 @@ class Formatter:
             val_unc_match_width: bool = None,
             bracket_unc_remove_seps: bool = None,
             unc_pm_whitespace: bool = None,
-            nan_include_exp: bool = None,
             val_unc_nan_include_exp: Union[bool, type(AutoValUncNanIncludeExp)] = None
     ):
         self.options = FormatOptions.make(
@@ -124,6 +124,7 @@ class Formatter:
             format_mode=format_mode,
             capital_exp_char=capital_exp_char,
             exp=exp,
+            nan_include_exp=nan_include_exp,
             use_prefix=use_prefix,
             extra_si_prefixes=extra_si_prefixes,
             extra_iec_prefixes=extra_iec_prefixes,
@@ -133,7 +134,6 @@ class Formatter:
             val_unc_match_widths=val_unc_match_width,
             bracket_unc_remove_seps=bracket_unc_remove_seps,
             unc_pm_whitespace=unc_pm_whitespace,
-            nan_include_exp=nan_include_exp,
             val_unc_nan_include_exp=val_unc_nan_include_exp
         )
 
@@ -159,6 +159,7 @@ class Formatter:
                    format_mode=options.format_mode,
                    capital_exp_char=options.capital_exp_char,
                    exp=options.exp,
+                   nan_include_exp=options.nan_include_exp,
                    use_prefix=options.use_prefix,
                    extra_si_prefixes=options.extra_si_prefixes,
                    extra_iec_prefixes=options.extra_iec_prefixes,
@@ -166,7 +167,6 @@ class Formatter:
                    val_unc_match_width=options.val_unc_match_widths,
                    bracket_unc_remove_seps=options.bracket_unc_remove_seps,
                    unc_pm_whitespace=options.unc_pm_whitespace,
-                   nan_include_exp=options.nan_include_exp,
                    val_unc_nan_include_exp=options.val_unc_nan_include_exp)
 
     @classmethod
