### 2023-06-16 c9f053a
nan_include_exp order
