### 2023-06-16 a3e0ec7
remove unc_nan_inf_exp options
