### 2023-06-16 6a5de76
nan_inf_exp tests
