### 2023-06-16 e635ca9
nan_inf_exp changes
