### 2023-06-16 40bfe17
rename capital_exp_char -> capitalize
