+++ docs/source/usage.rst
@@ -74,7 +74,7 @@ The global default settings can be viewed using
  'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,
  'precision': <class 'sciform.modes.AutoPrec'>,
  'format_mode': <FormatMode.FIXEDPOINT: 'fixed_point'>,
- 'capital_exp_char': False,
+ 'capitalize': False,
  'exp': <class 'sciform.modes.AutoExp'>,
  'nan_inf_exp': False,
  'use_prefix': False,
@@ -109,7 +109,7 @@ applied to setting global default settings.
  'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,
  'precision': 4,
  'format_mode': <FormatMode.ENGINEERING_SHIFTED: 'engineering_shifted'>,
- 'capital_exp_char': False,
+ 'capitalize': False,
  'exp': <class 'sciform.modes.AutoExp'>,
  'nan_inf_exp': False,
  'use_prefix': False,
