+++ docs/source/Formatting Options.rst
@@ -1,5 +1,6 @@
 .. _formatting_options:
 
+##################
 Formatting Options
 ##################
 
@@ -22,7 +23,7 @@ The different formatting modes control how ``mantissa``, ``base`` and
 ``exp`` are chosen for a given input float ``num``.
 
 Fixed Point
---------------
+-----------
 
 Fixed point notation is standard notation in which a float is displayed
 directly as a decimal number with no extra exponent.
@@ -336,3 +337,151 @@ Furthermore, just the ``c`` prefix can be included using the
 ...                   add_small_si_prefixes=True)
 >>> print(sform(25))
 2.5 da
+
+Include Exponent on nan and inf
+===============================
+
+Python supports ``float('nan')``, ``float('inf')``, and
+``float('-inf')``.
+Typically these are formatted to ``'nan'``, ``'inf'``, and ``'-inf'`` or
+``'NAN'``, ``'INF'``, and ``'-INF'`` respectively depending on
+``capitalize``.
+However, if ``nan_inf_exp=True`` (default ``False``), then, for
+scientific, engineering, and binary formatting modes, these will instead
+be formatted as, e.g. ``'(nan)e+00'``.
+
+>>> sform = Formatter(format_mode=FormatMode.SCIENTIFIC,
+...                   nan_inf_exp=True,
+...                   capitalize=True)
+>>> print(sform(float('-inf')))
+(-INF)E+00
+
+Value/Uncertainty Formatting
+============================
+
+For value/uncertainty formatting the value + uncertainty pair are
+formatted as follows.
+First, significant figure rounding is applied to the uncertainty
+according to the specified precision.
+Next the value is rounded to the same position as the uncertainty.
+The exponent is then determined using the format mode and the value.
+The value and the uncertainty are then formatted into a single string
+according to the options below.
+
+>>> sform = Formatter()
+>>> print(sform(123.456, 0.789))
+123.456 +/- 0.789
+
+Plus Minus Whitespace
+---------------------
+
+The user can enable (default) or disable white space around the plus/minus
+symbol when formatting value/uncertainties.
+
+>>> sform = Formatter()
+>>> print(sform(123.456, 0.789))
+123.456 +/- 0.789
+>>> sform = Formatter(unc_pm_whitespace=False)
+>>> print(sform(123.456, 0.789))
+123.456+/-0.789
+
+Bracket Uncertainty
+-------------------
+
+Instead of displaying ``123.456 +/- 0.789``, there is a notation where
+the uncertainty is shown in brackets after the value as
+``123.456(789)``.
+Here the ``(789)`` in parentheses is meant to be "matched up" with the
+finaly three digits of the value so that the 9 in the uncertainty is
+understood to appear in the thousandths place.
+This format is described in the
+`BIPM Guide Section 7.2.2 <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6#page=37>`_.
+We call this format "bracket uncertainty" mode.
+:mod:`sciform` provides this functionality via the ``bracket_unc``
+option:
+
+>>> sform = Formatter(bracket_unc=True)
+>>> print(sform(123.456, 0.789))
+123.456(789)
+
+Or with other options:
+
+>>> sform = Formatter(precision=2,
+...                   bracket_unc=True)
+>>> print(sform(123.456, 0.789))
+123.46(79)
+>>> sform = Formatter(precision=2,
+...                   format_mode=FormatMode.SCIENTIFIC,
+...                   bracket_unc=True)
+>>> print(sform(123.456, 0.789))
+(1.2346(79))e+02
+
+Remove Separators for Bracket Uncertainty
+--------------------------------------------
+
[…]
