+++ docs/source/usage.rst
@@ -48,6 +48,55 @@ formatted using the :mod:`sciform` :ref:`FSML <fsml>`.
 >>> print(f'{num:_!2f}')
 120_000
 
+Value/Uncertainty Formatting
+----------------------------
+
+One of, if not the, most important use cases for scientific formatting
+is formatting a value together with its specified uncertainty, e.g.
+``84.3 +/- 0.2``.
+:mod:`sciform` provides the ability to format pairs of floats into
+value/uncertainty strings.
+We attempt to follow
+`BIPM <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6>`_
+or `NIST <https://www.nist.gov/pml/nist-technical-note-1297>`_
+recomendations for conventions when possible.
+
+value/uncertainty pairs can be formatted either by passing two values
+into a :class:`Formatter`, and configuring the :class:`Formatter` using
+:ref:`formatting_options` and :ref:`val_unc_formatting_options`, or by
+using the :class:`vufloat` object.
+
+>>> val = 84.3
+>>> unc = 0.2
+>>> sform = Formatter(precision=2)
+>>> print(sform(val, unc))
+84.30 +/- 0.20
+>>> from sciform import vufloat
+>>> val_unc = vufloat(val, unc)
+>>> print(f'{val_unc:!2}')
+84.30 +/- 0.20
+
+Value/uncertainty pairs can also be shown in a common format where the
+uncertainty is displayed in parentheses after the value
+
+>>> print(f'{val_unc:!2S}')
+84.30(20)
+
+value/uncertainty pairs are formatted according to the following
+algorithm:
+
+#. Rounding is always performed using significant figure rounding
+   applied to the uncertainty. If a ``precision`` is supplied then the
+   uncertainty is rounded to that many signficant figures. Otherwise it
+   is not rounded.
+#. The value is rounded to the same digit place as the uncertainty
+#. The value for the exponent is resolved by applying the
+   ``format_mode`` to the value.
+#. The value and uncertainties mantissas are determined according to the
+   value of the exponent determined in the previous step.
+#. The value and uncertainty mantissas are formatted together with the
+   exponent according to the user-selected display options.
+
 Global Configuration
 ====================
 
