### 2023-06-17 4b2ed1b
Formatter docstring
