### 2023-06-17 2e0691c
val/unc example in formatter docstring
