### 2023-06-17 877e207
Catch bracket uncertainty issue when using non-default separators
