### 2023-06-17 b66cb34
sfloat and vufloat docstrings
