+++ tests/test_fsml.py
@@ -332,7 +332,6 @@ class TestFormatting(unittest.TestCase):
         self.do_test_case_dict(cases_dict)
 
     def test_non_finite(self):
-        # TODO: These behaviors need to be documented
         cases_dict = {
             float('nan'): {
                 '': 'nan',
@@ -356,7 +355,6 @@ class TestFormatting(unittest.TestCase):
         self.do_test_case_dict(cases_dict)
 
     def test_non_finite_with_exp(self):
-        # TODO: These behaviors need to be documented
         cases_dict = {
             float('nan'): {
                 '': 'nan',
