+++ tests/test_fsml.py
@@ -269,6 +269,30 @@ class TestFormatting(unittest.TestCase):
 
         self.do_test_case_dict(cases_dict)
 
+    def test_exp(self):
+        cases_dict: dict[float, dict[str, str]] = {
+            123.456: {
+                'ex-4': '1234560e-04',
+                'ex-3': '123456e-03',
+                'ex-2': '12345.6e-02',
+                'ex-1': '1234.56e-01',
+                'ex+0': '123.456e+00',
+                'ex+1': '12.3456e+01',
+                'ex+2': '1.23456e+02',
+                'ex+3': '0.123456e+03',
+                'ex+4': '0.0123456e+04',
+                'ex1': '12.3456e+01',
+                'ex2': '1.23456e+02',
+                'ex3': '0.123456e+03',
+                'ex4': '0.0123456e+04',
+                'rx0': '123.456e+00',
+                'rx-3': '123456e-03',
+                'rx+3': '0.123456e+03'
+            }
+        }
+
+        self.do_test_case_dict(cases_dict)
+
     def test_rounding(self):
         cases_dict: dict[float, dict[str, str]] = {
             99.99: {
