### 2023-06-17 510ce03
fit plot table example
