+++ docs/source/examples.rst
@@ -1,0 +1,267 @@
+Examples
+########
+
+Plotting and Tabulating Fit Data
+================================
+
+We are given 3 data sets which share the same x data
+
+::
+
+    x = [-0.0001, -7.777777777777778e-05, -5.555555555555556e-05, -3.3333333333333335e-05, -1.1111111111111112e-05, 1.1111111111111112e-05, 3.3333333333333335e-05, 5.555555555555556e-05, 7.777777777777778e-05, 0.0001]
+    y0 = [1000028705.1993496, 1000032562.3072652, 1000001896.3076739, 1000015911.7068863, 1000010552.1178038, 1000024250.5937256, 1000078382.654146, 1000099609.1405739, 1000156104.3810261, 1000218228.6977944]
+    y0_err = [10074.736165669336, 8935.559150444922, 9783.419334631266, 9603.772217328124, 8958.492409007613, 9691.25932834366, 9465.18285498962, 9252.428221037011, 10506.00188280341, 9674.67988710319]
+    y1 = [1000101738.923473, 1000075626.403057, 1000047095.4124904, 1000035342.0517923, 1000034627.8667482, 1000024117.4097912, 1000032427.2038687, 1000058361.2708515, 1000090132.8138337, 1000137137.590938]
+    y1_err = [9546.933912908422, 9865.10610577111, 11233.801341794, 8709.471433431938, 9260.554239646424, 11037.621922605267, 11397.162303260564, 10037.634586482105, 10076.884695349665, 9877.999777816845]
+    y2 = [1000162524.5101302, 1000112400.2588283, 1000095882.1735442, 1000055167.7660033, 1000026965.91884, 1000019054.4158406, 1000025166.114796, 1000035728.0662737, 1000075468.031305, 1000105069.4047513]
+    y2_err = [10476.46500331082, 10086.821923665966, 9559.819918015002, 10775.7594031851, 11296.833796893989, 10205.878907138671, 10025.395431977211, 10091.52840469254, 11880.794221192386, 10936.463066427263]
+
+
+We would like to perform quadratic fits to these data sets, visualize
+the results, and print the best fit parameters including the uncertainty
+reported by the fit routine.
+For these tasks we will require the ``numpy``, ``scipy``,
+``matplotlib``, and ``tabulate`` packages.
+Without ``sciform`` we can perform the fit and plot the data and best
+fit lines and print out a table of best fit parameters and
+uncertainties::
+
+    import numpy as np
+    import matplotlib.pyplot as plt
+    from scipy.optimize import curve_fit
+    from tabulate import tabulate
+
+    x = [-0.0001, -7.777777777777778e-05, -5.555555555555556e-05, -3.3333333333333335e-05, -1.1111111111111112e-05, 1.1111111111111112e-05, 3.3333333333333335e-05, 5.555555555555556e-05, 7.777777777777778e-05, 0.0001]
+    y0 = [1000028705.1993496, 1000032562.3072652, 1000001896.3076739, 1000015911.7068863, 1000010552.1178038, 1000024250.5937256, 1000078382.654146, 1000099609.1405739, 1000156104.3810261, 1000218228.6977944]
+    y0_err = [10074.736165669336, 8935.559150444922, 9783.419334631266, 9603.772217328124, 8958.492409007613, 9691.25932834366, 9465.18285498962, 9252.428221037011, 10506.00188280341, 9674.67988710319]
+    y1 = [1000101738.923473, 1000075626.403057, 1000047095.4124904, 1000035342.0517923, 1000034627.8667482, 1000024117.4097912, 1000032427.2038687, 1000058361.2708515, 1000090132.8138337, 1000137137.590938]
[…]
