### 2023-06-17 #114 @jagerber48
@NickleDave

I've updated the documentation to include my prototypical use case: https://sciform.readthedocs.io/en/stable/examples.html. Here I am doing two visualization tasks. I have x, y data which I am fitting an extracting best fit parameters for. The first visualization task is plotting the data. The second visualization task is displaying the best fit parameters (and their uncertainties from the fit routine) in a table.

`sciform` helps with the first plotting task by making it relatively straight forward (though with some admittedly not 100% straightforward helper functions) to convert the tick labels into SI prefix format.

`sciform` helps with the second table task by making it easy to format value/uncertainty pairs together for easy reading and order of magnitude comparison.

I imagine `sciform` will typically be used in python scripts or notebooks after some data analysis has been done, and now the user want to print analysis results to the terminal or notebook output. However, the result could also be saved into some sort of human-readable, text-based report which lives in memory or which is saved to the disk.

Instead of using `sciform` immediately at the conclusion of analysis, users could also use `sciform` while traversing a non-human readable data file to generate a rounded, human-readable version or summary of that data file. For example if the data file contains numeric or value/uncertainty type data.

I imagine adding an option to format strings into a "pretty" format using unicode characters and also a "latex" format similar to the `uncertainties` and other float formatting packages I linked above. Especially the "latex" format will open up more use cases for plotting (matplotlib requires latex for some formatting tasks) and report generation.
