+++ src/sciform/format_options.py
@@ -6,7 +6,7 @@ from pprint import pprint
 
 from sciform.modes import (FillMode, SignMode, GroupingSeparator,
                            UpperGroupingSeparators, LowerGroupingSeparators,
-                           DecimalGroupingSeparators, RoundMode, FormatMode,
+                           DecimalGroupingSeparators, RoundMode, ExpMode,
                            AutoExp, AutoPrec)
 
 
@@ -20,9 +20,9 @@ class FormatOptions:
     lower_separator: LowerGroupingSeparators
     round_mode: RoundMode
     precision: Union[int, type(AutoPrec)]
-    format_mode: FormatMode  # TODO: rename to exp_mode
-    capitalize: bool
+    exp_mode: ExpMode  # TODO: rename to exp_mode
     exp: Union[int, type(AutoExp)]
+    capitalize: bool
     nan_inf_exp: bool
     use_prefix: bool
     extra_si_prefixes: dict[int, str]
@@ -40,23 +40,23 @@ class FormatOptions:
                                      f'rounding, not {self.precision}.')
 
         if self.exp is not AutoExp:
-            if (self.format_mode is FormatMode.FIXEDPOINT
-                    or self.format_mode is FormatMode.PERCENT):
+            if (self.exp_mode is ExpMode.FIXEDPOINT
+                    or self.exp_mode is ExpMode.PERCENT):
                 if self.exp != 0:
                     raise ValueError(f'Exponent must must be 0, not '
                                      f'exp={self.exp}, for fixed point and '
-                                     f'percent format modes.')
-            elif (self.format_mode is FormatMode.ENGINEERING
-                    or self.format_mode is FormatMode.ENGINEERING_SHIFTED):
+                                     f'percent exponent modes.')
+            elif (self.exp_mode is ExpMode.ENGINEERING
+                  or self.exp_mode is ExpMode.ENGINEERING_SHIFTED):
                 if self.exp % 3 != 0:
                     raise ValueError(f'Exponent must be a multiple of 3, not '
-                                     f'exp={self.exp}, for engineering format '
-                                     f'modes.')
-            elif self.format_mode is FormatMode.BINARY_IEC:
+                                     f'exp={self.exp}, for engineering '
+                                     f'exponent modes.')
+            elif self.exp_mode is ExpMode.BINARY_IEC:
                 if self.exp % 10 != 0:
                     raise ValueError(f'Exponent must be a multiple of 10, not '
-                                     f'exp={self.exp}, for binary IEC format'
-                                     f'mode.')
+                                     f'exp={self.exp}, for binary IEC '
+                                     f'exponent mode.')
 
         if self.upper_separator not in get_args(UpperGroupingSeparators):
             raise ValueError(f'upper_separator {self.upper_separator} not in '
@@ -87,9 +87,9 @@ class FormatOptions:
             lower_separator: LowerGroupingSeparators = None,
             round_mode: RoundMode = None,
             precision: Union[int, type(AutoPrec)] = None,
-            format_mode: FormatMode = None,
-            capitalize: bool = None,
+            exp_mode: ExpMode = None,
             exp: Union[int, type(AutoExp)] = None,
+            capitalize: bool = None,
             nan_inf_exp: bool = None,
             use_prefix: bool = None,
             extra_si_prefixes: dict[int, str] = None,
@@ -120,12 +120,14 @@ class FormatOptions:
             round_mode = defaults.round_mode
         if precision is None:
             precision = defaults.precision
-        if format_mode is None:
[…]
