### 2023-06-17 4da58c8
Make percent an options, not an exponent mode
