+++ src/sciform/modes.py
@@ -100,9 +100,6 @@ class ExpMode(Enum):
     #: Fixed point
     FIXEDPOINT = 'fixed_point'
 
-    #: Percentage
-    PERCENT = 'percent'  # TODO: This should be an option, not an exp mode.
-
     #: Scientific
     SCIENTIFIC = 'scientific'
 
