+++ docs/source/usage.rst
@@ -124,6 +124,7 @@ The global default settings can be viewed using
  'exp_mode': <ExpMode.FIXEDPOINT: 'fixed_point'>,
  'exp': <class 'sciform.modes.AutoExp'>,
  'capitalize': False,
+ 'percent': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},
@@ -159,6 +160,7 @@ applied to setting global default settings.
  'exp_mode': <ExpMode.ENGINEERING_SHIFTED: 'engineering_shifted'>,
  'exp': <class 'sciform.modes.AutoExp'>,
  'capitalize': False,
+ 'percent': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},

[…]
