### 2023-06-18 dd0b7bb
fix percent mode bug (multiplying by 100 twice)
