+++ tests/test_val_unc_fsml.py
@@ -152,6 +152,18 @@ class TestFormatting(unittest.TestCase):
 
         self.do_unc_val_test_case_dict(cases_dict)
 
+    def test_percent(self):
+        cases_dict = {
+            (0.12462, 0.0001):
+                {
+                    '%': '(12.46 +/- 0.01)%',
+                    '%S': '(12.46(1))%'
+                }
+        }
+
+        self.do_unc_val_test_case_dict(cases_dict)
+
+
     def test_bracket_unc(self):
         cases_dict = {
             (123.456, 0.789):
