### 2023-06-18 200b99d
version bump to 0.13.2
- Replace FormatMode with ExpMode
- Make percent an options, not an exponent mode
  - updates and bugfixes for percent mode.
