+++ src/sciform/formatting.py
@@ -240,10 +240,14 @@ def format_val_unc(val: float, unc: float, options: FormatOptions):
         exp_str = unc_match.group('exp_str')
 
     if not options.bracket_unc:
-        if options.unc_pm_whitespace:
-            val_unc_str = f'{val_str} +/- {unc_str}'
+        if not options.unicode_pm:
+            pm_symb = '+/-'
         else:
-            val_unc_str = f'{val_str}+/-{unc_str}'
+            pm_symb = '±'
+        if options.unc_pm_whitespace:
+            val_unc_str = f'{val_str} {pm_symb} {unc_str}'
+        else:
+            val_unc_str = f'{val_str}{pm_symb}{unc_str}'
     else:
         unc_str = unc_str.lstrip('0.,_ ')
         if options.bracket_unc_remove_seps:
