### 2023-06-18 d7e18c0
Add superscript_exp feature and docs. Add pypi link
