+++ docs/source/usage.rst
@@ -125,6 +125,7 @@ The global default settings can be viewed using
  'exp': <class 'sciform.modes.AutoExp'>,
  'capitalize': False,
  'percent': False,
+ 'superscript_exp': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},
@@ -162,6 +163,7 @@ applied to setting global default settings.
  'exp': <class 'sciform.modes.AutoExp'>,
  'capitalize': False,
  'percent': False,
+ 'superscript_exp': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},
