### 2023-06-18 9596571
tests, superscript exp bug
