### 2023-06-18 e397020
Fix a bug involving space padding and separators
