### 2023-06-18 bdbb9c6
version bump to 0.15.2, rename changelog, update changelog
