### 2023-06-18 ce0032a
fix a bug that can miss top separator
