+++ tests/test_fsml.py
@@ -434,6 +434,20 @@ class TestFormatting(unittest.TestCase):
                 '.,_': '12.345,543_21',
                 '_,_': '12_345,543_21',
                 's,s': '12 345,543 21',
+            },
+            1234567.7654321: {
+                '': '1234567.7654321',
+                ',': '1,234,567.7654321',
+                ',.s': '1,234,567.765 432 1',
+                ',._': '1,234,567.765_432_1',
+                '_._': '1_234_567.765_432_1',
+                's.s': '1 234 567.765 432 1',
+                'n,n': '1234567,7654321',
+                '.,': '1.234.567,7654321',
+                '.,s': '1.234.567,765 432 1',
+                '.,_': '1.234.567,765_432_1',
+                '_,_': '1_234_567,765_432_1',
+                's,s': '1 234 567,765 432 1',
             }
         }
 
