### 2023-06-18 8bd0257
Add latex option
