+++ docs/source/usage.rst
@@ -126,6 +126,7 @@ The global default settings can be viewed using
  'capitalize': False,
  'percent': False,
  'superscript_exp': False,
+ 'latex': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},
@@ -164,6 +165,7 @@ applied to setting global default settings.
  'capitalize': False,
  'percent': False,
  'superscript_exp': False,
+ 'latex': False,
  'nan_inf_exp': False,
  'use_prefix': False,
  'extra_si_prefixes': {},
