### 2023-06-18 e0fd87e
Add latex feature and tests
