+++ tests/test_val_unc_formatter.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sciform import Formatter, ExpMode
+from sciform import Formatter, ExpMode, GroupingSeparator
 
 
 class TestFormatting(unittest.TestCase):
@@ -79,6 +79,25 @@ class TestFormatting(unittest.TestCase):
             (789, 0.01): {
                 Formatter(exp_mode=ExpMode.SCIENTIFIC,
                           superscript_exp=True): '(7.8900 +/- 0.0001)×10²'
+            }
+        }
+
+        self.do_test_case_dict(cases_dict)
+
+    def test_latex(self):
+        cases_dict = {
+            (12345, 0.2): {
+                Formatter(exp_mode=ExpMode.SCIENTIFIC,
+                          exp=-1,
+                          upper_separator=GroupingSeparator.UNDERSCORE,
+                          latex=True): r'\left(123\_450 \pm 2\right)\times 10^{-1}',
+
+                # Latex mode takes precedence over unicode_pm
+                Formatter(exp_mode=ExpMode.SCIENTIFIC,
+                          exp=-1,
+                          upper_separator=GroupingSeparator.UNDERSCORE,
+                          unicode_pm=True,
+                          latex=True): r'\left(123\_450 \pm 2\right)\times 10^{-1}'
             }
         }
 
