### 2023-06-18 dbbc4a2
latex mode docstring
