### 2023-06-18 6dc5ba0
refactor exp_str conversion
