### 2023-06-18 924cd3e
latex format documentation
