### 2023-06-18 fb4fdec
nan % fixes
