+++ CHANGELOG.rst
@@ -3,6 +3,7 @@
 
 * Add ``latex`` option with documentation and tests
 * Refactor exponent string conversion
+* Percent mode for non-finite numbers
 
 0.15.2 (2023-06-18)
 -------------------
