### 2023-06-18 b1930a9
more latex mode documentation
