+++ docs/source/usage.rst
@@ -129,6 +129,7 @@ The global default settings can be viewed using
  'latex': False,
  'nan_inf_exp': False,
  'use_prefix': False,
+ 'pp_prefix': False,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'bracket_unc': False,
@@ -168,6 +169,7 @@ applied to setting global default settings.
  'latex': False,
  'nan_inf_exp': False,
  'use_prefix': False,
+ 'pp_prefix': False,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'bracket_unc': False,
