+++ src/sciform/formatter.py
@@ -148,8 +148,10 @@ class Formatter:
             pp_prefix: bool = None,
             extra_si_prefixes: dict[int, str] = None,
             extra_iec_prefixes: dict[int, str] = None,
+            extra_pp_prefixes: dict[int, str] = None,
             add_c_prefix: bool = False,
             add_small_si_prefixes: bool = False,
+            add_ppth_prefix: bool = False,
             bracket_unc: bool = None,
             val_unc_match_widths: bool = None,
             bracket_unc_remove_seps: bool = None,
@@ -177,8 +179,10 @@ class Formatter:
             pp_prefix=pp_prefix,
             extra_si_prefixes=extra_si_prefixes,
             extra_iec_prefixes=extra_iec_prefixes,
+            extra_pp_prefixes=extra_pp_prefixes,
             add_c_prefix=add_c_prefix,
             add_small_si_prefixes=add_small_si_prefixes,
+            add_ppth_prefix=add_ppth_prefix,
             bracket_unc=bracket_unc,
             val_unc_match_widths=val_unc_match_widths,
             bracket_unc_remove_seps=bracket_unc_remove_seps,
@@ -216,6 +220,7 @@ class Formatter:
                    pp_prefix=options.pp_prefix,
                    extra_si_prefixes=options.extra_si_prefixes,
                    extra_iec_prefixes=options.extra_iec_prefixes,
+                   extra_pp_prefixes=options.extra_pp_prefixes,
                    bracket_unc=options.bracket_unc,
                    val_unc_match_widths=options.val_unc_match_widths,
                    bracket_unc_remove_seps=options.bracket_unc_remove_seps,
