+++ docs/source/usage.rst
@@ -132,6 +132,7 @@ The global default settings can be viewed using
  'pp_prefix': False,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
+ 'extra_pp_prefixes': {},
  'bracket_unc': False,
  'val_unc_match_widths': False,
  'bracket_unc_remove_seps': False,
@@ -172,6 +173,7 @@ applied to setting global default settings.
  'pp_prefix': False,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
+ 'extra_pp_prefixes': {},
  'bracket_unc': False,
  'val_unc_match_widths': False,
  'bracket_unc_remove_seps': False,
