### 2023-06-18 8c4528b
expose global ppth configuration
