+++ src/sciform/__init__.py
@@ -5,7 +5,8 @@ from sciform.float_formatting import sfloat, vufloat
 from sciform.format_options import (
     set_global_defaults, reset_global_defaults,
     global_add_c_prefix, global_add_small_si_prefixes,
-    global_reset_si_prefixes, global_reset_iec_prefixes,
+    global_add_ppth_prefix, global_reset_si_prefixes,
+    global_reset_iec_prefixes, global_reset_pp_prefixes,
     print_global_defaults, GlobalDefaultsContext)
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
                            ExpMode)
@@ -13,6 +14,7 @@ from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
 __all__ = ['__version__', 'Formatter', 'sfloat', 'set_global_defaults',
            'reset_global_defaults', 'global_add_c_prefix',
            'global_add_small_si_prefixes', 'global_reset_si_prefixes',
-           'global_reset_iec_prefixes', 'print_global_defaults',
+           'global_add_ppth_prefix', 'global_reset_iec_prefixes',
+           'global_reset_pp_prefixes', 'print_global_defaults',
            'GlobalDefaultsContext', 'FillMode', 'SignMode',
            'GroupingSeparator', 'RoundMode', 'ExpMode', 'vufloat']
