+++ src/sciform/format_options.py
@@ -27,11 +27,11 @@ class FormatOptions:
     superscript_exp: bool
     latex: bool
     nan_inf_exp: bool
-    use_prefix: bool
-    pp_prefix: bool
+    prefix_exp: bool
+    parts_per_exp: bool
     extra_si_prefixes: dict[int, str]
     extra_iec_prefixes: dict[int, str]
-    extra_pp_prefixes: dict[int, str]
+    extra_parts_per_forms: dict[int, str]
     bracket_unc: bool
     val_unc_match_widths: bool
     bracket_unc_remove_seps: bool
@@ -103,14 +103,14 @@ class FormatOptions:
             superscript_exp: bool = None,
             latex: bool = None,
             nan_inf_exp: bool = None,
-            use_prefix: bool = None,
-            pp_prefix: bool = None,
+            prefix_exp: bool = None,
+            parts_per_exp: bool = None,
             extra_si_prefixes: dict[int, str] = None,
             extra_iec_prefixes: dict[int, str] = None,
-            extra_pp_prefixes: dict[int, str] = None,
+            extra_parts_per_forms: dict[int, str] = None,
             add_c_prefix: bool = False,
             add_small_si_prefixes: bool = False,
-            add_ppth_prefix: bool = False,
+            add_ppth_form: bool = False,
             bracket_unc: bool = None,
             val_unc_match_widths: bool = None,
             bracket_unc_remove_seps: bool = None,
@@ -150,10 +150,10 @@ class FormatOptions:
             latex = defaults.latex
         if nan_inf_exp is None:
             nan_inf_exp = defaults.nan_inf_exp
-        if use_prefix is None:
-            use_prefix = defaults.use_prefix
-        if pp_prefix is None:
-            pp_prefix = defaults.pp_prefix
+        if prefix_exp is None:
+            prefix_exp = defaults.prefix_exp
+        if parts_per_exp is None:
+            parts_per_exp = defaults.parts_per_exp
         if extra_si_prefixes is None:
             extra_si_prefixes = copy(defaults.extra_si_prefixes)
         else:
@@ -162,10 +162,10 @@ class FormatOptions:
             extra_iec_prefixes = copy(defaults.extra_iec_prefixes)
         else:
             extra_iec_prefixes = copy(extra_iec_prefixes)
-        if extra_pp_prefixes is None:
-            extra_pp_prefixes = copy(defaults.extra_pp_prefixes)
+        if extra_parts_per_forms is None:
+            extra_parts_per_forms = copy(defaults.extra_parts_per_forms)
         else:
-            extra_pp_prefixes = copy(extra_pp_prefixes)
+            extra_parts_per_forms = copy(extra_parts_per_forms)
         if bracket_unc is None:
             bracket_unc = defaults.bracket_unc
         if val_unc_match_widths is None:
@@ -190,9 +190,9 @@ class FormatOptions:
             if +2 not in extra_si_prefixes:
                 extra_si_prefixes[+2] = 'h'
 
-        if add_ppth_prefix:
-            if -3 not in extra_pp_prefixes:
-                extra_pp_prefixes[-3] = 'ppth'
+        if add_ppth_form:
+            if -3 not in extra_parts_per_forms:
+                extra_parts_per_forms[-3] = 'ppth'
 
         return cls(
             fill_mode=fill_mode,
@@ -210,11 +210,11 @@ class FormatOptions:
             superscript_exp=superscript_exp,
             latex=latex,
             nan_inf_exp=nan_inf_exp,
-            use_prefix=use_prefix,
-            pp_prefix=pp_prefix,
+            prefix_exp=prefix_exp,
[…]
