+++ src/sciform/formatter.py
@@ -98,9 +98,9 @@ class Formatter:
       such as ``float('nan')`` or ``float('inf')`` should be formatted
       with exponent symbols when exponent modes including exponent
       symbols are selected.
-    :param use_prefix: ``bool`` indicating if exponents should be
+    :param prefix_exp: ``bool`` indicating if exponents should be
       replaced with either SI or IEC prefixes as appropriate.
-    :param pp_prefix: ``bool`` indicating if "parts-per" prefixes should
+    :param parts_per_exp: ``bool`` indicating if "parts-per" prefixes should
       be used.
     :param extra_si_prefixes: ``dict[int, str]`` mapping additional
       exponent values to si prefixes.
@@ -144,14 +144,14 @@ class Formatter:
             superscript_exp: bool = None,
             latex: bool = None,
             nan_inf_exp: bool = None,
-            use_prefix: bool = None,
-            pp_prefix: bool = None,
+            prefix_exp: bool = None,
+            parts_per_exp: bool = None,
             extra_si_prefixes: dict[int, str] = None,
             extra_iec_prefixes: dict[int, str] = None,
-            extra_pp_prefixes: dict[int, str] = None,
+            extra_parts_per_forms: dict[int, str] = None,
             add_c_prefix: bool = False,
             add_small_si_prefixes: bool = False,
-            add_ppth_prefix: bool = False,
+            add_ppth_form: bool = False,
             bracket_unc: bool = None,
             val_unc_match_widths: bool = None,
             bracket_unc_remove_seps: bool = None,
@@ -175,14 +175,14 @@ class Formatter:
             superscript_exp=superscript_exp,
             latex=latex,
             nan_inf_exp=nan_inf_exp,
-            use_prefix=use_prefix,
-            pp_prefix=pp_prefix,
+            prefix_exp=prefix_exp,
+            parts_per_exp=parts_per_exp,
             extra_si_prefixes=extra_si_prefixes,
             extra_iec_prefixes=extra_iec_prefixes,
-            extra_pp_prefixes=extra_pp_prefixes,
+            extra_parts_per_forms=extra_parts_per_forms,
             add_c_prefix=add_c_prefix,
             add_small_si_prefixes=add_small_si_prefixes,
-            add_ppth_prefix=add_ppth_prefix,
+            add_ppth_form=add_ppth_form,
             bracket_unc=bracket_unc,
             val_unc_match_widths=val_unc_match_widths,
             bracket_unc_remove_seps=bracket_unc_remove_seps,
@@ -216,11 +216,11 @@ class Formatter:
                    superscript_exp=options.superscript_exp,
                    latex=options.latex,
                    nan_inf_exp=options.nan_inf_exp,
-                   use_prefix=options.use_prefix,
-                   pp_prefix=options.pp_prefix,
+                   prefix_exp=options.prefix_exp,
+                   parts_per_exp=options.parts_per_exp,
                    extra_si_prefixes=options.extra_si_prefixes,
                    extra_iec_prefixes=options.extra_iec_prefixes,
-                   extra_pp_prefixes=options.extra_pp_prefixes,
+                   extra_parts_per_forms=options.extra_parts_per_forms,
                    bracket_unc=options.bracket_unc,
                    val_unc_match_widths=options.val_unc_match_widths,
                    bracket_unc_remove_seps=options.bracket_unc_remove_seps,
