+++ tests/test_float_formatting.py
@@ -71,7 +71,7 @@ class TestFormatting(unittest.TestCase):
                           latex=True): r'123\_450\times 10^{-1}',
                 Formatter(exp_mode=ExpMode.SCIENTIFIC,
                           exp=3,
-                          use_prefix=True,
+                          prefix_exp=True,
                           latex=True): r'12.345 k'
             }
         }
@@ -88,22 +88,22 @@ class TestFormatting(unittest.TestCase):
         }
         self.do_test_case_dict(cases_dict)
 
-    def test_pp_prefix(self):
+    def test_parts_per_exp(self):
         cases_dict = {
             123e-3: {
                 Formatter(exp_mode=ExpMode.SCIENTIFIC,
                           exp=-3,
-                          use_prefix=True,
-                          pp_prefix=True,
-                          add_ppth_prefix=True): '123 ppth',
+                          prefix_exp=True,
+                          parts_per_exp=True,
+                          add_ppth_form=True): '123 ppth',
                 Formatter(exp_mode=ExpMode.SCIENTIFIC,
                           exp=-6,
-                          use_prefix=True,
-                          pp_prefix=True): '123000 ppm',
+                          prefix_exp=True,
+                          parts_per_exp=True): '123000 ppm',
                 Formatter(exp_mode=ExpMode.SCIENTIFIC,
                           exp=-2,
-                          use_prefix=True,
-                          pp_prefix=True): '12.3e-02'
+                          prefix_exp=True,
+                          parts_per_exp=True): '12.3e-02'
             }
         }
 

[…]
