+++ docs/source/fsml.rst
@@ -3,6 +3,9 @@
 Format Specification Mini Language
 ##################################
 
+.. module:: sciform
+   :noindex:
+
 Specification
 =============
 
