### 2023-06-19 425d418
fix doctests
