### 2023-06-19 057ef97
usage global config documentation
