### 2023-06-19 2f4e808
move Extra Exponent Replacements to be subsection under Exponent String Replacement
