+++ CHANGELOG.rst
@@ -1,17 +1,17 @@
 0.17.0 (2023-06-??)
 -------------------
 
-* Add parts-per notation feature, documentation, and tests
-* **[BREAKING]** Rename ``use_prefix`` option to ``prefix_exp``
-* Fix typos in binary IEC prefixes table
-* Fix some cross links in documentation
+* Add parts-per notation feature, documentation, and tests.
+* **[BREAKING]** Rename ``use_prefix`` option to ``prefix_exp``.
+* Fix typos in binary IEC prefixes table.
+* Fix some cross links in documentation.
 
 0.16.0 (2023-06-18)
 -------------------
 
-* Add ``latex`` option with documentation and tests
-* Refactor exponent string conversion
-* Percent mode for non-finite numbers
+* Add ``latex`` option with documentation and tests.
+* Refactor exponent string conversion.
+* Percent mode for non-finite numbers.
 
 0.15.2 (2023-06-18)
 -------------------
@@ -33,5 +33,5 @@
 0.14.0 (2023-06-17)
 -------------------
 
-* Add Changelog
+* Add Changelog.
 * Add ``unicode_pm`` option with documentation and tests.
