+++ tests/test_config.py
@@ -48,4 +48,3 @@ class TestConfig(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
-
