+++ tests/test_val_unc_fsml.py
@@ -298,4 +298,4 @@ class TestFormatting(unittest.TestCase):
 
 
 if __name__ == '__main__':
-    unittest.main(verbosity=2)
+    unittest.main()
