+++ tests/test_float_formatting.py
@@ -105,3 +105,7 @@ class TestFormatting(unittest.TestCase):
         }
 
         self.do_test_case_dict(cases_dict)
+
+
+if __name__ == '__main__':
+    unittest.main()
