+++ tests/test_val_unc_fsml.py
@@ -297,5 +297,5 @@ class TestFormatting(unittest.TestCase):
         self.do_unc_val_test_case_dict(cases_dict)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
