+++ tests/test_val_unc_formatter.py
@@ -126,3 +126,7 @@ class TestFormatting(unittest.TestCase):
         }
 
         self.do_test_case_dict(cases_dict)
+
+
+if __name__ == "__main__":
+    unittest.main()
