### 2023-06-19 79764e2
Refactor formatting and format_utils
