+++ src/sciform/format_utils.py
@@ -10,24 +10,16 @@ from sciform.prefix import (si_val_to_prefix_dict, iec_val_to_prefix_dict,
                             pp_val_to_prefix_dict)
 
 
-def get_top_digit(num: float) -> int:
+def get_top_digit(num: float, binary=False) -> int:
     if not isfinite(num):
         return 0
-    num = abs(num)
-    int_part = int(num)
-    if int_part == 0:
-        magnitude = 1
+    elif num == 0:
+        return 0
     else:
-        magnitude = int(log10(int_part)) + 1
-    max_digits = sys.float_info.dig
-    if magnitude >= max_digits:
-        return magnitude
-
-    try:
-        top_digit = floor(log10(num))
-    except ValueError:
-        top_digit = 0
-    return top_digit
+        if not binary:
+            return floor(log10(abs(num)))
+        else:
+            return floor(log2(abs(num)))
 
 
 def get_bottom_digit(num: float) -> int:
@@ -47,7 +39,7 @@ def get_bottom_digit(num: float) -> int:
 
     frac_part = num - int_part
     multiplier = 10 ** (max_digits - magnitude)
-    frac_digits = multiplier + int(multiplier * frac_part + 0.5)
+    frac_digits = multiplier + floor(multiplier * frac_part + 0.5)
     while frac_digits % 10 == 0:
         frac_digits /= 10
     precision = int(log10(frac_digits))
@@ -65,53 +57,58 @@ def get_mantissa_exp_base(
         num: float,
         exp_mode: ExpMode,
         exp: Union[int, type(AutoExp)] = None) -> (float, int, int):
-    if num == 0:
-        if exp is AutoExp:
-            exp = 0
-        base = 10
-    elif exp_mode is ExpMode.FIXEDPOINT:
-        if exp is not AutoExp:
-            if exp != 0:
-                warn('Attempt to set exponent explicity in fixed point '
-                     'exponent mode. Coercing exponent to 0.')
-        exp = 0
-        base = 10
-    elif exp_mode is ExpMode.SCIENTIFIC:
-        if exp is AutoExp:
-            exp = floor(log10(abs(num)))
-        base = 10
-    elif (exp_mode is ExpMode.ENGINEERING
-          or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-        if exp is not AutoExp:
-            if exp % 3 != 0:
-                warn(f'Attempt to set exponent explicity to a non-integer '
-                     f'multiple of 3 in engineering mode. Coercing to the '
-                     f'next lower multiple of 3.')
-                exp = (exp // 3) * 3
-            if exp_mode is ExpMode.ENGINEERING_SHIFTED:
-                warn(f'Engineering shifted mode is ignored when setting '
-                     f'exponent explicitly.')
-        else:
-            exp = floor(log10(abs(num)))
-            if exp_mode is ExpMode.ENGINEERING:
-                exp = (exp // 3) * 3
-            else:
-                exp = ((exp + 1) // 3) * 3
-        base = 10
-    elif (exp_mode is ExpMode.BINARY
-          or exp_mode is ExpMode.BINARY_IEC):
-        if exp is AutoExp:
-            exp = floor(log2(abs(num)))
-            if exp_mode is ExpMode.BINARY_IEC:
-                exp = (exp // 10) * 10
-        elif exp_mode is ExpMode.BINARY_IEC:
-            warn(f'Binary IEC mode is ignored when setting exponent '
-                 f'explicitly')
+    if (exp_mode is ExpMode.BINARY
+            or exp_mode is ExpMode.BINARY_IEC):
         base = 2
     else:
-        raise ValueError(f'Unhandled exponent mode: {exp_mode}')
+        base = 10
 
-    mantissa = num * base**-exp
+    if not isfinite(num):
+        mantissa = num
[…]
