+++ src/sciform/formatting.py
@@ -6,14 +6,62 @@ from sciform.modes import FillMode, ExpMode, SignMode, AutoExp
 from sciform.format_options import FormatOptions, RoundMode
 from sciform.format_utils import (get_mantissa_exp_base, get_exp_str,
                                   get_top_digit,
-                                  get_top_and_bottom_digit,
                                   get_round_digit,
                                   format_float_by_top_bottom_dig,
-                                  convert_exp_str)
+                                  convert_exp_str,
+                                  latex_translate)
 from sciform.grouping import add_separators
 
 
 # TODO: ppm format
+
+
+def format_non_inf(num: float, options: FormatOptions) -> str:
+    if isfinite(num):
+        raise ValueError(f'format_non_inf() cannot format finite float {num}.')
+
+    if options.nan_inf_exp:
+        exp_mode = options.exp_mode
+
+        exp = options.exp
+        if options.exp is AutoExp:
+            exp = 0
+
+        if exp_mode is ExpMode.FIXEDPOINT:
+            exp_str = ''
+        elif (exp_mode is ExpMode.SCIENTIFIC
+              or exp_mode is ExpMode.ENGINEERING
+              or exp_mode is ExpMode.ENGINEERING_SHIFTED):
+            exp_str = f'e+{exp:02d}'
+        else:
+            exp_str = f'b+{exp:02d}'
+    else:
+        exp_str = ''
+    exp_str = convert_exp_str(exp_str,
+                              options.prefix_exp,
+                              options.parts_per_exp,
+                              options.latex,
+                              options.superscript_exp,
+                              options.extra_si_prefixes,
+                              options.extra_iec_prefixes,
+                              options.extra_parts_per_forms)
+
+    if exp_str != '':
+        result = f'({num}){exp_str}'
+    else:
+        result = f'{num}'
+        if options.percent:
+            result = f'({result})%'
+
+    if options.capitalize:
+        result = result.upper()
+    else:
+        result = result.lower()
+
+    if options.latex:
+        result = latex_translate(result)
+
+    return result
 
 
 def format_float(num: float, options: FormatOptions) -> str:
@@ -22,62 +70,16 @@ def format_float(num: float, options: FormatOptions) -> str:
     precision = options.precision
     top_padded_digit = options.top_dig_place
     sign_mode = options.sign_mode
-    capitalize = options.capitalize
-    fill_char = FillMode.to_char(options.fill_mode)
+
     if not isfinite(num):
-        if options.nan_inf_exp:
-            if options.exp is AutoExp:
-                exp = 0
-            else:
-                exp = options.exp
-            if exp_mode is ExpMode.FIXEDPOINT:
-                exp_str = ''
-            elif (exp_mode is ExpMode.SCIENTIFIC
-                  or exp_mode is ExpMode.ENGINEERING
-                  or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-                exp_str = f'e+{exp:02d}'
-            else:
-                exp_str = f'b+{exp:02d}'
-        else:
-            exp_str = ''
-        exp_str = convert_exp_str(exp_str,
-                                  options.prefix_exp,
-                                  options.latex,
-                                  options.superscript_exp,
-                                  options.parts_per_exp,
-                                  options.extra_si_prefixes,
-                                  options.extra_iec_prefixes,
-                                  options.extra_parts_per_forms)
-
-        if exp_str != '':
-            full_str = f'({num}){exp_str}'
-        else:
-            full_str = f'{num}'
-            if options.percent:
-                full_str = f'({full_str})%'
-
-        if capitalize:
[…]
