### 2023-06-19 bf56780
Cleanup
- Some minor code restructuring
- Minor documentation cleanup
- More verbose comments to explain trickier parts of the formatting algorithm.
