+++ src/sciform/formatting.py
@@ -13,9 +13,6 @@ from sciform.format_utils import (get_mantissa_exp_base, get_exp_str,
 from sciform.grouping import add_separators
 
 
-# TODO: ppm format
-
-
 def format_non_inf(num: float, options: FormatOptions) -> str:
     if isfinite(num):
         raise ValueError(f'format_non_inf() cannot format finite float {num}.')
@@ -35,23 +32,25 @@ def format_non_inf(num: float, options: FormatOptions) -> str:
             exp_str = f'e+{exp:02d}'
         else:
             exp_str = f'b+{exp:02d}'
+
+        exp_str = convert_exp_str(exp_str,
+                                  options.prefix_exp,
+                                  options.parts_per_exp,
+                                  options.latex,
+                                  options.superscript_exp,
+                                  options.extra_si_prefixes,
+                                  options.extra_iec_prefixes,
+                                  options.extra_parts_per_forms)
     else:
         exp_str = ''
-    exp_str = convert_exp_str(exp_str,
-                              options.prefix_exp,
-                              options.parts_per_exp,
-                              options.latex,
-                              options.superscript_exp,
-                              options.extra_si_prefixes,
-                              options.extra_iec_prefixes,
-                              options.extra_parts_per_forms)
 
     if exp_str != '':
         result = f'({num}){exp_str}'
     else:
         result = f'{num}'
-        if options.percent:
-            result = f'({result})%'
+
+    if options.percent:
+        result = f'({result})%'
 
     if options.capitalize:
         result = result.upper()
@@ -92,7 +91,12 @@ def format_float(num: float, options: FormatOptions) -> str:
     mantissa_rounded = round(mantissa, -round_digit)
 
     if mantissa_rounded == 0:
-        # This catches an edge case involving negative precision
+        '''
+        This catches an edge case involving negative precision when the
+        resulting mantissa is zero after the second rounding. This 
+        result is technically correct (e.g. 0e+03 = 0e+00), but sciform 
+        always presents zero values with an exponent of zero.
+        '''
         exp = 0
 
     if mantissa_rounded == -0.0:
@@ -105,6 +109,15 @@ def format_float(num: float, options: FormatOptions) -> str:
                                                   sign_mode,
                                                   fill_char)
 
+    upper_separator = options.upper_separator.to_char()
+    decimal_separator = options.decimal_separator.to_char()
+    lower_separator = options.lower_separator.to_char()
+    mantissa_str = add_separators(mantissa_str,
+                                  upper_separator,
+                                  decimal_separator,
+                                  lower_separator,
+                                  group_size=3)
+
     exp_str = get_exp_str(exp, exp_mode, options.capitalize)
     exp_str = convert_exp_str(exp_str,
                               options.prefix_exp,
@@ -115,19 +128,10 @@ def format_float(num: float, options: FormatOptions) -> str:
                               options.extra_iec_prefixes,
                               options.extra_parts_per_forms)
 
-    upper_separator = options.upper_separator.to_char()
-    decimal_separator = options.decimal_separator.to_char()
-    lower_separator = options.lower_separator.to_char()
-    mantissa_str = add_separators(mantissa_str,
-                                  upper_separator,
-                                  decimal_separator,
-                                  lower_separator,
-                                  group_size=3)
-
     result = f'{mantissa_str}{exp_str}'
 
     if options.percent:
[…]
