+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.17.0"
+__version__ = "0.17.1"
 
 # TODO: PDG sig figs
 # TODO: get exponent based on larger of val/unc
