### 2023-06-19 b81dc37
Change FillMode.to_char()
