+++ src/sciform/modes.py
@@ -1,4 +1,4 @@
-from typing import get_args, Literal
+from typing import Literal
 from enum import Enum
 
 
@@ -8,10 +8,6 @@ class AutoExp:
 
 class AutoPrec:
     pass
-
-
-def option_warn_str(value, options: type(Literal)):
-    return f'Flag \'{value}\' not in {get_args(options)}.'
 
 
 class FillMode(Enum):
