### 2023-06-19 ee3525b
Change some unit tests and add others
