+++ src/sciform/__init__.py
@@ -11,6 +11,8 @@ from sciform.format_options import (
     global_add_ppth_form, global_reset_si_prefixes,
     global_reset_iec_prefixes, global_reset_parts_per_forms,
     print_global_defaults, GlobalDefaultsContext)
+
+# TODO: Add AutoExp and AutoPrec (and documentation)
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
                            ExpMode)
 
