### 2023-06-20 770e32f
add AutoExp and AutoPrec into public API and add lots of documentation
