+++ src/sciform/__init__.py
@@ -14,12 +14,13 @@ from sciform.format_options import (
 
 # TODO: Add AutoExp and AutoPrec (and documentation)
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
-                           ExpMode)
+                           ExpMode, AutoExp, AutoPrec)
 
-__all__ = ['__version__', 'Formatter', 'sfloat', 'set_global_defaults',
-           'reset_global_defaults', 'global_add_c_prefix',
-           'global_add_small_si_prefixes', 'global_reset_si_prefixes',
-           'global_add_ppth_form', 'global_reset_iec_prefixes',
-           'global_reset_parts_per_forms', 'print_global_defaults',
-           'GlobalDefaultsContext', 'FillMode', 'SignMode',
-           'GroupingSeparator', 'RoundMode', 'ExpMode', 'vufloat']
+__all__ = ['__version__', 'Formatter', 'sfloat', 'vufloat',
+           'set_global_defaults', 'reset_global_defaults',
+           'global_add_c_prefix', 'global_add_small_si_prefixes',
+           'global_reset_si_prefixes', 'global_add_ppth_form',
+           'global_reset_iec_prefixes', 'global_reset_parts_per_forms',
+           'print_global_defaults', 'GlobalDefaultsContext', 'FillMode',
+           'SignMode', 'GroupingSeparator', 'RoundMode', 'ExpMode', 'AutoExp',
+           'AutoPrec']
