+++ docs/source/project.rst
@@ -5,3 +5,13 @@ Changelog
 =========
 
 .. include:: ../../CHANGELOG.rst
+
+
+Roadmap
+=======
+
+* Support for
+  `decimal <https://docs.python.org/3/library/decimal.html#module-decimal>`_
+  objects in addition to floats. This is imporant for supporting
+  formatting numbers with a large (e.g. >15) number of significant
+  digits or a high precision.
