### 2023-06-20 6a6a86e
better pdg examples
