### 2023-06-20 cb2bbd1
slight bracket_unc_remove_seps code reorg
