### 2023-06-21 c916076
cleanup options.rst
