+++ docs/source/options.rst
@@ -47,6 +47,12 @@ mantissa ``m`` satisfies ``1 <= m < 10``.
 >>> print(sform(123.456))
 1.23456e+02
 
+Note that for all exponent modes, the exponent integer is always displayed
+with a sign symbol (+ or -) and is left padded with a zero so that it is
+at least two digits wide. There are no options to modify this behavior
+for standard exponent display. The :ref:`superscript_exp` or
+:ref:`latex_format` can be used as alternatives.
+
 Engineering Notation
 --------------------
 
@@ -107,8 +113,6 @@ notation in base-2.
 >>> sform = Formatter(exp_mode=ExpMode.BINARY_IEC)
 >>> print(sform(2048))
 2b+10
-
-Here ``b`` exponent symbol indicates base-2 instead of base-10.
 
 Fixed Exponent
 ==============
@@ -189,8 +193,10 @@ Furthermore, just the ``c`` prefix can be included using the
 >>> print(sform(25))
 2.5 da
 
-A non-standard parts-per-thousand form, ``ppth``, can be accessed with
+A parts-per-thousand form, ``ppth``, can be accessed with
 the ``add_ppth_form`` option.
+Note that ``ppth`` is not a standard notation for "parts-per-thousand",
+but it is one that the author has found useful.
 
 >>> sform = Formatter(exp_mode=ExpMode.ENGINEERING,
 ...                   parts_per_exp=True,
@@ -210,7 +216,8 @@ In both cases, the rounding applies to the mantissa determined after
 identifying the appropriate exponent for display based on the selected
 exponent mode.
 In some cases, the rounding results in a modification to the chosen
-exponent.
+exponent (e.g. presenting 9.99 with a precision of 1 in scientific
+exponent mode).
 This is taken into account before the final presentation.
 
 In both cases, if no explicit precision value or number of significant
@@ -255,7 +262,7 @@ Precision simply indicates the number of digits to be displayed past the
 decimal point.
 So, e.g., a precision of 2 indicates rounding to the hundredths, or
 10\ :sup:`-2`, place.
-Most of the built-in string formatting mini-language is based on
+Much of the Python built-in string formatting mini-language is based on
 precision presentation.
 
 >>> from sciform import RoundMode
@@ -288,8 +295,8 @@ decimal symbol.
 Additionally, :mod:`sciform` also supports including separation characters
 between groups of three digits both above the decimal symbol and below
 the decimal symbols.
-No separator, ``','``, ``'.'``, ``' '``, ``'_'`` can all be used as
-"upper" separator characters and no separator, ``' '``, and ``'_'`` can
+``''``, ``','``, ``'.'``, ``' '``, ``'_'`` can all be used as
+"upper" separator characters and ``''``, ``' '``, and ``'_'`` can
 all be used as "lower" separator characters.
 Note that the upper separator character must be different than the
 decimal separator.
@@ -359,7 +366,7 @@ Left Filling
 ============
 
 The :ref:`rounding` options described above can be used to control how
-many digits to the left of either the most-significant digit or the
+many digits to the right of either the most-significant digit or the
 decimal point are displayed.
 It is also possible, using "fill" options, to add digits to the left of
 the most-significant digit.
@@ -392,6 +399,8 @@ appended to the end of the formatted string.
 >>> print(sform(0.12345, 0.001))
 (12.345 +/- 0.100)%
 
+.. _superscript_exp:
+
 Superscript Exponent Format
 ===========================
 
@@ -402,6 +411,8 @@ standard superscript notation as opposed to e.g. ``e+02`` notation.
 ...                   superscript_exp=True)
 >>> print(sform(789))
 7.89×10²
+
+.. _latex_format:
 
 Latex Format
 ============
[…]
