### 2023-06-21 add6ff5
cleanup fsml.rst
