+++ docs/source/fsml.rst
@@ -10,12 +10,12 @@ Specification
 =============
 
 :mod:`sciform` :ref:`formatting options <formatting_options>` can be
-applied to the formatting of :class:`sfloat` objects by using string
-formatting analogous to the built-in formatting of :class:`float`
-objects.
+applied to the formatting of :class:`sfloat` or :class:`vufloat` objects
+by using string formatting analogous to the built-in formatting of
+:class:`float` objects.
 The :mod:`sciform` format specification mini language is given by::
 
-    format_spec        ::=  [fill "="][sign]["#"][fill_top_digit][upper_separator][decimal_separator][lower_separator][round_mode precision][exp_mode]["x" exp][prefix_mode]
+    format_spec        ::=  [fill "="][sign]["#"][fill_top_digit][upper_separator][decimal_separator][lower_separator][round_mode precision][exp_mode]["x" exp][p][S]
 
     fill               ::=  "0" | " "
     sign               ::=  "+" | "-" | " "
@@ -24,10 +24,9 @@ The :mod:`sciform` format specification mini language is given by::
     decimal_separator  ::=  "." | ","
     lower_separator    ::=  "n" | "s" | "_"
     round_mode         ::=  "!" | "."
-    prec               ::=  -?digit+
+    precision          ::=  [+-]?digit+
     exp_mode           ::=  "f" | "F" | "%" | "e" | "E" | "r" | "R" | "b" | "B" |
-    exp                ::=  [+-]digit+
-    prefix_mode        ::=  p
+    exp                ::=  [+-]?digit+
 
 Details about the terms in the FSML are described below.
 
@@ -52,11 +51,9 @@ Details about the terms in the FSML are described below.
        | (``'#'``)
      - Alternate mode is enabled (disabled by default) if the ``'#'``
        flag is included in the format specification. In engineering
-       notation (``r`` or ``R``), the alternate mode coerces the
-       mantissa `m` to be ``0.1 <= m < 100`` rather than
-       ``1 <= m < 1000``. In binary mode (``b`` or ``B``), the alternate
-       flag coerces the mantissa to be between ``1 <= m < 1024`` rather
-       than ``1 <= m < 2``.
+       mode (``r`` or ``R``), the alternate mode engages
+       :ref:`engineering_shifted` mode. In binary mode (``b`` or ``B``),
+       the alternate mode engages :ref:`binary_iec` mode.
    * - | fill_top_digit
        | (``\d+``)
      - Any non-negative integer, default (0). Indicates the decimal or
@@ -87,12 +84,11 @@ Details about the terms in the FSML are described below.
    * - | round_mode
        | (``'!'``, ``'.'``)
      - Indicates whether the float will be rounded and displayed
-       according to precision (digits past the decimal point) or
-       significant figure. ``'.'`` indicates precision mode and ``'!'``
-       indicates significant figure mode. E.g.
-       ``f'{sfloat(123.456):.2f}'`` gives ``'123.46'`` while
-       ``f'{sfloat(123.456):!2f}'`` gives ``'120'``.
-   * - | prec
+       according to precision ``'.'`` (digits past the decimal point) or
+       significant figures ``'!'``. E.g. ``f'{sfloat(123.456):.2f}'``
+       gives ``'123.46'`` while ``f'{sfloat(123.456):!2f}'`` gives
+       ``'120'``.
+   * - | precision
        | (``[+-]?\d+``)
      - Integer indicating the precision or number of significant figures
        to which the float shall be rounded and displayed. Can be
@@ -111,24 +107,26 @@ Details about the terms in the FSML are described below.
        capitalization of the exponent symbol matches the capitalization
        of the exponent mode flag.
 
-       * ``'f'`` and ``'F'`` indicate fixed point mode in which no
[…]
