### 2023-06-22 46184b0
update examples.rst
