+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.18.0"
+__version__ = "0.18.1"
 
 from sciform.formatter import Formatter
 from sciform.float_formatting import sfloat, vufloat
