+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+0.18.1 (2023-06-21)
+
+* Documentation improvements
+
 0.18.0 (2023-06-19)
 -------------------
 
