### 2023-06-22 422855e
\text{} around prefix exponents for latex mode
