+++ tests/test_val_unc_fsml.py
@@ -29,7 +29,7 @@ class TestFormatting(unittest.TestCase):
                     '!2f': '123.46 +/- 0.79',
                     '!3f': '123.456 +/- 0.789',
                     '!4f': '123.4560 +/- 0.7890',
-                    '!2fS': '123.46(79)',
+                    '!2f()': '123.46(79)',
                 },
             (0.789, 123.456):
                 {
@@ -41,13 +41,13 @@ class TestFormatting(unittest.TestCase):
                     '!5f': '0.79 +/- 123.46',
                     '!6f': '0.789 +/- 123.456',
                     '!7f': '0.7890 +/- 123.4560',
-                    '!1fS': '0(100)',
-                    '!2fS': '0(120)',
-                    '!3fS': '1(123)',
-                    '!4fS': '0.8(123.5)',
-                    '!5fS': '0.79(123.46)',
-                    '!6fS': '0.789(123.456)',
-                    '!7fS': '0.7890(123.4560)',
+                    '!1f()': '0(100)',
+                    '!2f()': '0(120)',
+                    '!3f()': '1(123)',
+                    '!4f()': '0.8(123.5)',
+                    '!5f()': '0.79(123.46)',
+                    '!6f()': '0.789(123.456)',
+                    '!7f()': '0.7890(123.4560)',
                 }
         }
 
@@ -62,7 +62,7 @@ class TestFormatting(unittest.TestCase):
                     '!2e': '(1.2346 +/- 0.0079)e+02',
                     '!3e': '(1.23456 +/- 0.00789)e+02',
                     '!4e': '(1.234560 +/- 0.007890)e+02',
-                    '!2eS': '(1.2346(79))e+02',
+                    '!2e()': '(1.2346(79))e+02',
                 },
             (0.789, 123.456):
                 {
@@ -74,13 +74,13 @@ class TestFormatting(unittest.TestCase):
                     '!5e': '(0.0079 +/- 1.2346)e+02',
                     '!6e': '(0.00789 +/- 1.23456)e+02',
                     '!7e': '(0.007890 +/- 1.234560)e+02',
-                    '!1eS': '(0(1))e+02',
-                    '!2eS': '(0.0(1.2))e+02',
-                    '!3eS': '(0.01(1.23))e+02',
-                    '!4eS': '(0.008(1.235))e+02',
-                    '!5eS': '(0.0079(1.2346))e+02',
-                    '!6eS': '(0.00789(1.23456))e+02',
-                    '!7eS': '(0.007890(1.234560))e+02',
+                    '!1e()': '(0(1))e+02',
+                    '!2e()': '(0.0(1.2))e+02',
+                    '!3e()': '(0.01(1.23))e+02',
+                    '!4e()': '(0.008(1.235))e+02',
+                    '!5e()': '(0.0079(1.2346))e+02',
+                    '!6e()': '(0.00789(1.23456))e+02',
+                    '!7e()': '(0.007890(1.234560))e+02',
                 },
             (0, 123.456):
                 {
@@ -92,13 +92,13 @@ class TestFormatting(unittest.TestCase):
                     '!5e': '(0.0000 +/- 1.2346)e+02',
                     '!6e': '(0.00000 +/- 1.23456)e+02',
                     '!7e': '(0.000000 +/- 1.234560)e+02',
-                    '!1eS': '(0(1))e+02',
-                    '!2eS': '(0.0(1.2))e+02',
-                    '!3eS': '(0.00(1.23))e+02',
-                    '!4eS': '(0.000(1.235))e+02',
-                    '!5eS': '(0.0000(1.2346))e+02',
-                    '!6eS': '(0.00000(1.23456))e+02',
-                    '!7eS': '(0.000000(1.234560))e+02',
[…]
