### 2023-06-22 7bfa06e
try doctest
