+++ .github/workflows/python-package.yml
@@ -38,3 +38,6 @@ jobs:
     - name: Test with pytest
       run: |
         pytest
+    - name: Doc test
+      run: |
+        docs/make doctest
