+++ .github/workflows/python-package.yml
@@ -38,6 +38,6 @@ jobs:
     - name: Test with pytest
       run: |
         pytest
-    - name: Doc test
+    - name: Doctest
       run: |
-        docs/make doctest
+        pytest --doctest-glob="*.rst"
