+++ tests/test_fsml.py
@@ -249,7 +249,7 @@ class TestFormatting(unittest.TestCase):
                 '#r': '12.3456e+03',
                 '#.-3r': '0e+00',
                 '#.-2r': '0e+00',
-                '#.-1r': '10e+03', 
+                '#.-1r': '10e+03',
                 '#.0r': '12e+03',
                 '#.1r': '12.3e+03',
                 '#.2r': '12.35e+03',
