### 2023-06-22 b5b5253
Do doctest with other tests
