+++ .github/workflows/python-package.yml
@@ -37,7 +37,4 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        pytest
-    - name: Doctest
-      run: |
         pytest --doctest-glob="*.rst"
