### 2023-06-23 61454cc
Start refactoring tests to use lists instead of dicts
