### 2023-06-23 a913311
refactor float formatter and float fsml tests to use lists/tuples instead of dicts
