+++ tests/test_float_fsml.py
@@ -1,0 +1,592 @@
+import unittest
+
+from sciform import sfloat, GlobalDefaultsContext
+
+
+FloatFormatterCases = list[tuple[float, list[tuple[str, str]]]]
+
+
+class TestFormatting(unittest.TestCase):
+    def run_float_fsml_cases(self, cases_list: FloatFormatterCases):
+        for num, num_formats in cases_list:
+            for format_spec, expected_num_str in num_formats:
+                snum = sfloat(num)
+                snum_str = f'{snum:{format_spec}}'
+                with self.subTest(num=num, format_spec=format_spec,
+                                  expected_num_str=expected_num_str,
+                                  actual_num_str=snum_str):
+                    self.assertEqual(snum_str, expected_num_str)
+
+    def test_fixed_point(self):
+        cases_list = [
+            (123.456, [
+                ('f', '123.456'),
+                ('.-3f', '0'),
+                ('.-2f', '100'),
+                ('.-1f', '120'),
+                ('.0f', '123'),
+                ('.1f', '123.5'),
+                ('.2f', '123.46'),
+                ('.3f', '123.456'),
+                ('.4f', '123.4560'),
+                ('!1f', '100'),
+                ('!2f', '120'),
+                ('!3f', '123'),
+                ('!4f', '123.5'),
+                ('!5f', '123.46'),
+                ('!6f', '123.456'),
+                ('!7f', '123.4560')
+            ]),
+            (0.00062607, [
+                ('f', '0.00062607'),
+                ('.-1f', '0'),
+                ('.0f', '0'),
+                ('.1f', '0.0'),
+                ('.2f', '0.00'),
+                ('.3f', '0.001'),
+                ('.4f', '0.0006'),
+                ('.5f', '0.00063'),
+                ('.6f', '0.000626'),
+                ('.7f', '0.0006261'),
+                ('.8f', '0.00062607'),
+                ('.9f', '0.000626070'),
+                ('!1f', '0.0006'),
+                ('!2f', '0.00063'),
+                ('!3f', '0.000626'),
+                ('!4f', '0.0006261'),
+                ('!5f', '0.00062607'),
+                ('!6f', '0.000626070')
+            ])
+        ]
+
+        self.run_float_fsml_cases(cases_list)
+
+    def test_percent(self):
+        cases_list = [
+            (0.123456, [
+                ('%', '12.3456%'),
+                ('.-3%', '0%'),
+                ('.-2%', '0%'),
+                ('.-1%', '10%'),
+                ('.0%', '12%'),
+                ('.1%', '12.3%'),
+                ('.2%', '12.35%'),
+                ('.3%', '12.346%'),
+                ('.4%', '12.3456%'),
+                ('!1%', '10%'),
+                ('!2%', '12%'),
+                ('!3%', '12.3%'),
+                ('!4%', '12.35%'),
+                ('!5%', '12.346%'),
+                ('!6%', '12.3456%'),
+                ('!7%', '12.34560%')
+            ]),
+            (0.00062607, [
+                ('%', '0.062607%'),
+                ('.-1%', '0%'),
+                ('.0%', '0%'),
+                ('.1%', '0.1%'),
+                ('.2%', '0.06%'),
+                ('.3%', '0.063%'),
+                ('.4%', '0.0626%'),
+                ('.5%', '0.06261%'),
+                ('.6%', '0.062607%'),
+                ('.7%', '0.0626070%'),
+                ('.8%', '0.06260700%'),
+                ('.9%', '0.062607000%'),
[…]
