+++ tests/test_val_unc_formatter.py
@@ -3,145 +3,143 @@ import unittest
 from sciform import Formatter, ExpMode, GroupingSeparator
 
 
+ValUncFormatterCases = list[tuple[tuple[float, float],
+                                  list[tuple[Formatter, str]]]]
+
+
 class TestFormatting(unittest.TestCase):
-    def do_test_case_dict(self, cases_dict: dict[tuple[float, float],
-                                                 dict[Formatter, str]]):
-        for (val, unc), fmt_dict in cases_dict.items():
-            for formatter, expected_num_str in fmt_dict.items():
+    def run_val_unc_formatter_cases(self, cases_list: ValUncFormatterCases):
+        for (val, unc), formats_list in cases_list:
+            for formatter, expected_val_unc_str in formats_list:
                 snum_str = formatter(val, unc)
                 with self.subTest(val=val, unc=unc,
-                                  expected_num_str=expected_num_str,
+                                  expected_num_str=expected_val_unc_str,
                                   actual_num_str=snum_str):
-                    self.assertEqual(snum_str, expected_num_str)
+                    self.assertEqual(snum_str, expected_val_unc_str)
 
     def test_bracket_unc(self):
-        cases_dict = {
-            (123.456, 0.789):
-                {
-                    Formatter(bracket_unc=True): '123.456(789)',
-                    Formatter(exp_mode=ExpMode.SCIENTIFIC,
-                              bracket_unc=True): '(1.23456(789))e+02',
-                    Formatter(exp_mode=ExpMode.ENGINEERING,
-                              bracket_unc=True): '(123.456(789))e+00',
-                    Formatter(exp_mode=ExpMode.ENGINEERING_SHIFTED,
-                              bracket_unc=True): '(0.123456(789))e+03',
-                    Formatter(exp_mode=ExpMode.SCIENTIFIC,
-                              exp=+1,
-                              bracket_unc=True): '(12.3456(789))e+01',
-                    Formatter(exp_mode=ExpMode.SCIENTIFIC,
-                              exp=-1,
-                              bracket_unc=True): '(1234.56(7.89))e-01',
-                }
-        }
+        cases_list = [
+            ((123.456, 0.789), [
+                (Formatter(bracket_unc=True), '123.456(789)'),
+                (Formatter(exp_mode=ExpMode.SCIENTIFIC,
+                           bracket_unc=True), '(1.23456(789))e+02'),
+                (Formatter(exp_mode=ExpMode.ENGINEERING,
+                           bracket_unc=True), '(123.456(789))e+00'),
+                (Formatter(exp_mode=ExpMode.ENGINEERING_SHIFTED,
+                           bracket_unc=True), '(0.123456(789))e+03'),
+                (Formatter(exp_mode=ExpMode.SCIENTIFIC,
+                           exp=+1,
+                           bracket_unc=True), '(12.3456(789))e+01'),
+                (Formatter(exp_mode=ExpMode.SCIENTIFIC,
+                           exp=-1,
+                           bracket_unc=True), '(1234.56(7.89))e-01'),
+            ])
+        ]
 
-        self.do_test_case_dict(cases_dict)
+        self.run_val_unc_formatter_cases(cases_list)
 
     def test_percent(self):
-        cases_dict = {
-            (0.123_456_78, 0.000_002_55):
-                {
-                    Formatter(percent=True,
-                              lower_separator=GroupingSeparator.UNDERSCORE):
-                    '(12.345_678 +/- 0.000_255)%',
-                    Formatter(percent=True,
-                              bracket_unc=True,
-                              lower_separator=GroupingSeparator.UNDERSCORE):
-                    '(12.345_678(255))%'
[…]
