### 2023-06-23 8e616a4
refactor val unc fsml testing to lists/tuple instead of dicts
