+++ tests/test_val_unc_fsml.py
@@ -3,16 +3,16 @@ import unittest
 from sciform import vufloat, GlobalDefaultsContext
 
 
+ValUncFSMLCases = list[tuple[tuple[float, float],
+                             list[tuple[str, str]]]]
 NAN = float('nan')
 INF = float('inf')
 
 
 class TestFormatting(unittest.TestCase):
-    def do_unc_val_test_case_dict(
-            self,
-            cases_dict: dict[tuple[float, float], dict[str, str]]):
-        for (val, unc), fmt_dict in cases_dict.items():
-            for format_spec, expected_str in fmt_dict.items():
+    def run_val_unc_fsml_cases(self, cases_list: ValUncFSMLCases):
+        for (val, unc), formats_list in cases_list:
+            for format_spec, expected_str in formats_list:
                 vunum = vufloat(val, unc)
                 vunum_str = f'{vunum:{format_spec}}'
                 with self.subTest(val=val, unc=unc, format_spec=format_spec,
@@ -21,298 +21,283 @@ class TestFormatting(unittest.TestCase):
                     self.assertEqual(vunum_str, expected_str)
 
     def test_fixed(self):
-        cases_dict = {
-            (123.456, 0.789):
-                {
-                    'f': '123.456 +/- 0.789',
-                    '!1f': '123.5 +/- 0.8',
-                    '!2f': '123.46 +/- 0.79',
-                    '!3f': '123.456 +/- 0.789',
-                    '!4f': '123.4560 +/- 0.7890',
-                    '!2f()': '123.46(79)',
-                },
-            (0.789, 123.456):
-                {
-                    'f': '0.789 +/- 123.456',
-                    '!1f': '0 +/- 100',
-                    '!2f': '0 +/- 120',
-                    '!3f': '1 +/- 123',
-                    '!4f': '0.8 +/- 123.5',
-                    '!5f': '0.79 +/- 123.46',
-                    '!6f': '0.789 +/- 123.456',
-                    '!7f': '0.7890 +/- 123.4560',
-                    '!1f()': '0(100)',
-                    '!2f()': '0(120)',
-                    '!3f()': '1(123)',
-                    '!4f()': '0.8(123.5)',
-                    '!5f()': '0.79(123.46)',
-                    '!6f()': '0.789(123.456)',
-                    '!7f()': '0.7890(123.4560)',
-                }
-        }
+        cases_list = [
+            ((123.456, 0.789), [
+                ('f', '123.456 +/- 0.789'),
+                ('!1f', '123.5 +/- 0.8'),
+                ('!2f', '123.46 +/- 0.79'),
+                ('!3f', '123.456 +/- 0.789'),
+                ('!4f', '123.4560 +/- 0.7890'),
+                ('!2f()', '123.46(79)')
+            ]),
+            ((0.789, 123.456), [
+                ('f', '0.789 +/- 123.456'),
+                ('!1f', '0 +/- 100'),
+                ('!2f', '0 +/- 120'),
+                ('!3f', '1 +/- 123'),
+                ('!4f', '0.8 +/- 123.5'),
+                ('!5f', '0.79 +/- 123.46'),
+                ('!6f', '0.789 +/- 123.456'),
+                ('!7f', '0.7890 +/- 123.4560'),
+                ('!1f()', '0(100)'),
+                ('!2f()', '0(120)'),
+                ('!3f()', '1(123)'),
+                ('!4f()', '0.8(123.5)'),
+                ('!5f()', '0.79(123.46)'),
[…]
