### 2023-06-25 36bc52d
tests passing after converting inputs to decimals
