+++ src/sciform/format_utils.py
@@ -4,25 +4,42 @@ from math import floor, log10, log2, isfinite
 from warnings import warn
 import re
 from copy import copy
+from decimal import Decimal
 
 from sciform.modes import ExpMode, RoundMode, SignMode, AutoExp, AutoPrec
 from sciform.prefix import (si_val_to_prefix_dict, iec_val_to_prefix_dict,
                             pp_val_to_prefix_dict)
 
 
-def get_top_digit(num: float, binary=False) -> int:
+def get_top_digit(num: Decimal, binary=False) -> int:
     if not isfinite(num):
         return 0
-    elif num == 0:
-        return 0
+    if not binary:
+        _, digits, exp = num.as_tuple()
+        if exp == 'n':
+            return 0
+        return len(digits) + exp - 1
     else:
-        if not binary:
-            return floor(log10(abs(num)))
-        else:
-            return floor(log2(abs(num)))
+        return floor(log2(abs(num)))
+    # if not isfinite(num):
+    #     return 0
+    # elif num == 0:
+    #     return 0
+    # else:
+    #     if not binary:
+    #         return floor(log10(abs(num)))
+    #     else:
+    #         return floor(log2(abs(num)))
 
 
-def get_bottom_digit(num: float) -> int:
+def get_bottom_digit(num: Decimal) -> int:
+    if not isfinite(num):
+        return 0
+    if isinstance(num, Decimal):
+        round(num, 14)
+        _, digits, exp = num.as_tuple()
+        return exp
+
     if not isfinite(num):
         return 0
 
@@ -54,9 +71,11 @@ def get_top_and_bottom_digit(num: float) -> tuple[int, int]:
 
 
 def get_mantissa_exp_base(
-        num: float,
+        num: Decimal,
         exp_mode: ExpMode,
         exp: Union[int, type(AutoExp)] = None) -> (float, int, int):
+    num = Decimal(num)
+
     if (exp_mode is ExpMode.BINARY
             or exp_mode is ExpMode.BINARY_IEC):
         base = 2
@@ -68,7 +87,7 @@ def get_mantissa_exp_base(
         if exp is AutoExp:
             exp = 0
     elif num == 0:
-        mantissa = 0
+        mantissa = Decimal(0)
         if exp is AutoExp:
             exp = 0
     else:
@@ -107,14 +126,16 @@ def get_mantissa_exp_base(
                          'multiple of 10 in binary IEC mode. Coercing to the '
                          'next lower multiple of 10.')
                     exp = (exp // 10) * 10
-
+        base = Decimal(base)
+        # exp = Decimal(exp)
         mantissa = num * base**-exp
-
+        mantissa = Decimal(mantissa).normalize()
     return mantissa, exp, base
 
 
 def get_exp_str(exp: int, exp_mode: ExpMode,
                 capitalize: bool) -> str:
+    exp = int(exp)
     if exp_mode is exp_mode.FIXEDPOINT:
         exp_str = ''
     elif (exp_mode is ExpMode.SCIENTIFIC
@@ -131,7 +152,7 @@ def get_exp_str(exp: int, exp_mode: ExpMode,
     return exp_str
 
 
-def get_sign_str(num: float, sign_mode: SignMode) -> str:
+def get_sign_str(num: Decimal, sign_mode: SignMode) -> str:
     if num < 0:
         sign_str = '-'
     else:
@@ -146,7 +167,7 @@ def get_sign_str(num: float, sign_mode: SignMode) -> str:
     return sign_str
 
 
-def get_round_digit(num: float,
+def get_round_digit(num: Decimal,
                     round_mode: RoundMode,
                     precision: Union[int, type(AutoPrec)],
                     pdg_sig_figs: bool = False) -> int:
@@ -159,7 +180,7 @@ def get_round_digit(num: float,
             else:
[…]
