+++ src/sciform/formatting.py
@@ -1,6 +1,7 @@
 from math import isfinite
 from warnings import warn
 import re
+from decimal import Decimal, InvalidOperation
 
 from sciform.modes import ExpMode, SignMode, AutoExp
 from sciform.format_options import FormatOptions, RoundMode
@@ -13,7 +14,8 @@ from sciform.format_utils import (get_mantissa_exp_base, get_exp_str,
 from sciform.grouping import add_separators
 
 
-def format_non_inf(num: float, options: FormatOptions) -> str:
+def format_non_inf(num: Decimal, options: FormatOptions) -> str:
+    num = float(num)
     if isfinite(num):
         raise ValueError(f'format_non_inf() cannot format finite float {num}.')
 
@@ -63,7 +65,8 @@ def format_non_inf(num: float, options: FormatOptions) -> str:
     return result
 
 
-def format_float(num: float, options: FormatOptions) -> str:
+def format_num(num: Decimal, options: FormatOptions) -> str:
+    num = Decimal(num)
     if not isfinite(num):
         return format_non_inf(num, options)
 
@@ -76,7 +79,7 @@ def format_float(num: float, options: FormatOptions) -> str:
     precision = options.precision
     mantissa, temp_exp, base = get_mantissa_exp_base(num, exp_mode, exp)
     round_digit = get_round_digit(mantissa, round_mode, precision)
-    mantissa_rounded = round(mantissa, -round_digit)
+    mantissa_rounded = Decimal(round(mantissa, -int(round_digit)))
 
     '''
     Repeat mantissa + exponent discovery after rounding in case rounding
@@ -85,7 +88,7 @@ def format_float(num: float, options: FormatOptions) -> str:
     rounded_num = mantissa_rounded * base**temp_exp
     mantissa, exp, base = get_mantissa_exp_base(rounded_num, exp_mode, exp)
     round_digit = get_round_digit(mantissa, round_mode, precision)
-    mantissa_rounded = round(mantissa, -round_digit)
+    mantissa_rounded = Decimal(round(mantissa, -int(round_digit)))
 
     if mantissa_rounded == 0:
         '''
@@ -100,7 +103,7 @@ def format_float(num: float, options: FormatOptions) -> str:
         mantissa_rounded = abs(mantissa_rounded)
 
     fill_char = options.fill_mode.to_char()
-    mantissa_str = format_float_by_top_bottom_dig(mantissa_rounded,
+    mantissa_str = format_float_by_top_bottom_dig(mantissa_rounded.normalize(),
                                                   options.top_dig_place,
                                                   round_digit,
                                                   options.sign_mode,
@@ -136,7 +139,7 @@ def format_float(num: float, options: FormatOptions) -> str:
     return result
 
 
-def format_val_unc(val: float, unc: float, options: FormatOptions):
+def format_val_unc(val: Decimal, unc: Decimal, options: FormatOptions):
     if options.round_mode is RoundMode.PREC:
         warn('Precision round mode not available for value/uncertainty '
              'formatting. Rounding is always applied as significant figures '
@@ -147,6 +150,8 @@ def format_val_unc(val: float, unc: float, options: FormatOptions):
     if options.percent:
         val *= 100
         unc *= 100
+        val = val.normalize()
+        unc = unc.normalize()
 
     # Find the digit place to round to
     if isfinite(unc) and unc != 0:
@@ -156,9 +161,18 @@ def format_val_unc(val: float, unc: float, options: FormatOptions):
 
     round_digit = get_round_digit(round_driver, RoundMode.SIG_FIG,
                                   options.precision, options.pdg_sig_figs)
[…]
