### 2023-06-26 9a3d77a
More Decimal friendly re-factoring
