### 2023-06-26 #114 @NickleDave
Hi @jagerber48 thank you for your patience--we wanted to get input from other community members about whether this package was in scope.

Thank you also for updating the documentation with a use case. That is exactly the kind of concrete example that really helps users understand what you are trying to do for them.

We have decided that, yes, we will proceed with a review.

Please go ahead and make a full submission. Be sure to mention this issue by number when you do so ("as discussed in #114") and please be sure to complete the pre-review survey when you do make the submission. Appreciate it!

Once you have opened that issue referencing this one, I will close this. We will then put out a call for an editor and reviewers.
