### 2023-06-28 0c161d1
add back float arithmetic and docstring. Cleanup some methods
