### 2023-07-18 e520654
remove unnecessary cast to int
