### 2023-07-21 #114 @jagerber48
@NickleDave I've made the full submission at https://github.com/pyOpenSci/software-submission/issues/121.

One question before next steps: I have a few high level and lower design questions about the package. Some are about the overall architecture of the code and some are about "should I include this feature or this requirement". I'm curious if these types of questions are in-scope for the code review. Or if the code review should be thought of as reviewing the quality of the code and giving general advice based on the code at one snapshot in time (at one version number). I may as well mention some specific questions I have here and then you can better inform me about their appropriateness for discussion. These are the questions I have that I'm not sure are in scope for review. I also have some questions that I'm more sure are in scope for review (like should I add more unit tests, how can I improve continuous integration).

- All of the formatting options are captured as data fields on the `FormatOptions` object. However, these options need to be repeated in full many times throughout the code in function signatures and bodies for a few reasons. What can be done to mitigate this repetition? Specifically, this repetition means that a lot of (somewhat error-prone) work needs to be done if I ever want to add new options.
- How broad should the scope of the package be? Should it JUST be responsible for formatting individual numbers or pairs of numbers? Or should it support some obvious possible utility functions such as (1) performing arithmetical operations on `sfloat` (sciform formattable float objects) or (2) formatting lists of numbers? Should `sciform` get into more involved formatting involving units?
  - For example, one nice way to map the `sciform` functions over sequences or arrays is using `np.vectorize`. But it is worth making `sciform` depend on `numpy` for this?
- If an `sfloat` class is supported should an `sDecimal` class be supported?
- Right now the default plus/minus value/uncertainty separator is `"+/-"`. But given unicode is prolific now, should the default be `"±"`?
