### 2023-07-23 a14b6d7
Use explicit isfinite checks instead of InvalidOperation
