### 2023-07-23 a6bea3d
remove unnecessary int cast
