### 2023-07-23 73c6309
Rename sfloat -> SciNum, vufloat -> SciNumUnc, remove arithmetic operations
