+++ tests/test_val_unc_fsml.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sciform import vufloat, GlobalDefaultsContext
+from sciform import SciNumUnc, GlobalDefaultsContext
 
 
 ValUncFSMLCases = list[tuple[tuple[float, float],
@@ -13,7 +13,7 @@ class TestFormatting(unittest.TestCase):
     def run_val_unc_fsml_cases(self, cases_list: ValUncFSMLCases):
         for (val, unc), formats_list in cases_list:
             for format_spec, expected_str in formats_list:
-                vunum = vufloat(val, unc)
+                vunum = SciNumUnc(val, unc)
                 vunum_str = f'{vunum:{format_spec}}'
                 with self.subTest(val=val, unc=unc, format_spec=format_spec,
                                   expected_str=expected_str,
