### 2023-07-23 23e92ac
Documentation cleanup float-centric language
