### 2023-07-24 5a873d1
Add section on decimals and floats
