+++ docs/source/usage.rst
@@ -243,3 +243,114 @@ In particular, not all avaible options can be accessed using the
 :ref:`FSML <fsml>`, so the only way to modify these options while using
 :class:`SciNum` or :class:`SciNumUnc` formatting is via the global
 configuration settings.
+
+Note on Decimals and Floats
+===========================
+
+Numerical data can be stored in Python
+`float <https://docs.python.org/3/library/functions.html#float>`_
+or
+`Decimal <https://docs.python.org/3/library/decimal.html>`_ objects.
+:class:`floats` represent numbers using binary which means
+:class:`floats` are often only approximations of the decimal numbers
+users have in mind when they use :class:`floats`.
+By contrast, :class:`Decimal` objects store a string of integers
+representing the decimal digits of the represented number so
+:class:`Decimal` objects are, therefore, exact representations of
+decimal numbers.
+
+Both of these representations have finite precision which can cause
+unexpected issues when manipulating numerical data.
+However, the :class:`Decimal` class is much better suited to address
+these issues.
+Internally, the :mod:`sciform` module uses :class:`Decimal`
+representations of the numbers it is formatting.
+
+Here I would like to highlight some important facts and possible issues
+with :class:`float` objects that users should be aware of if they are
+concerned with the exact decimal representation of their numerical data.
+
+* Python uses
+  `double-precision floating-point format <https://en.wikipedia.org/wiki/Double-precision_floating-point_format>`_
+  for its :class:`float`. In this format a :class:`float` occupies 64
+  bit of memory: 52 bits for the mantissa, 11 bits for the exponent and
+  one bit for the sign.
+* Any decimal with 15 digits between about ``+/- 1.8e+308`` can be
+  uniquely represented by a :class:`float`. However, two decimals with
+  more than 15 digits may map to the same :class:`float`. For example,
+  ``float(8.000000000000001) == float(8.000000000000002)`` returns
+  ``True``.
+* If any :class:`float` is converted to a decimal with at least 17
+  digits then it will be converted back to the same float. See
+  `decimal precision of binary floating point number <https://www.exploringbinary.com/decimal-precision-of-binary-floating-point-numbers/>`_.
+  However, many :class:`floats` can be "round-tripped" many fewer
+  digits. The `__repr__` for the python :class:`float` class converts
+  the float to a decimal representation with the minimum number of
+  digits such that it round trips to the same :class:`float`. For
+  example we can see the exact decimal representation of the
+  :class:`float` which ``0.1`` is mapped to:
+  ``print(Decimal(float(0.1)))`` gives
+  ``0.1000000000000000055511151231257827021181583404541015625``.
+  However ``print(float(0.1))`` just gives ``0.1``. That is,
+  ``0.1000000000000000055511151231257827021181583404541015625`` and
+  ``0.1`` map to the same :class:`float` but the :class:`float`
+  ``__repr__`` algorithim presents us with the shorter (more readable)
+  decimal representation.
+
+The above considerations on :class:`floats` can lead to two surprising
+issues that I will highlight.
+
+#. Rounding. Consider the decimal numbers ``0.0355`` and ``0.00355``.
+   If we round these to two significant figures using a "round-to-even"
+   strategy we expect these to round to ``0.036`` and ``0.0036``
+   respectively. However, if we try to perform this rounding for
+   :class:`floats` we get an unexpected result. We see that
+   ``round(0.00355, 4)`` gives ``0.0036`` as expected but
+   ``round(0.0355, 3)`` gives ``0.035``. We can see the issue by looking
+   at the decimal representations of the corresponding :class:`floats`.
+   ``print(Decimal(0.0355))`` gives
[…]
