### 2023-07-24 a5a64ae
Remove traces of floats in internal algorithms
