### 2023-07-24 0b748b9
for consistency, use same strategy to get top digit for decimal and binary formats.
