+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.21.0"
+__version__ = "0.22.0"
 
 from sciform.formatter import Formatter
 from sciform.scinum import SciNum, SciNumUnc
