### 2023-07-24 #114 @NickleDave
Hi @jagerber48, happy to help.

These are all good questions to ask yourself as a developer, and I have definitely found myself pondering similar questions before.

However, I can't give you a detailed answer here, because I would feel like I'm starting to review.

In fact, some of these questions start to be about scope, and ideally we should not run a review just for the purpose of figuring out scope. That's something that should be determined ahead of time.

We do want to help you though.
Let's do the following in this case:
- We'll hold off on starting the review for now. We will still review, but let's make sure you have answers to at least some of these questions first.
- Please ask your questions above in our discourse forum, using the [coding help category](https://pyopensci.discourse.group/c/coding-help/10). You can ask however you feel is appropriate, but my guess is that it you'll get the most out of it if you identify the questions that are most crucial to figure out, and create a separate "topic" (post) for each one. Someone who is unfamiliar with the package will need a lot of context just to understand one question, so you are more likely to get a helpful answer if you can present each question clearly, with a concise title and as much context as possible.

A related practice that I find helpful is to keep a "dev diary". I write down questions like this each day I do dev work, and I also prioritize my to-dos. If the same questions or ideas keep popping up, then it helps me know that I really need to prioritize working on them. I also include links to other code, papers, etc., that give me concrete examples--if I can't find anyone else who is doing what I have in mind, then that tells me something.

Hope that's somewhat helpful--I'm only telling you because I wish I had gotten into this practice much sooner, along with using project management tools like [GitHub Projects](https://docs.github.com/en/issues/planning-and-tracking-with-projects/learning-about-projects/about-projects).

Please ask these questions on our forum and let's take it from there. Let's time box that process--say, two weeks max--and then we'll start the review.
