### 2023-07-25 #114 @NickleDave
That's perfect, thank you @jagerber48.

The question on the forum is very well stated and I think you will get good feedback.

I think you are exactly right to take a more conservative approach for now. One thing I see happen is that developers get excited about adding new features and solving the related programming problems. There's nothing wrong with that, of course. (It's one of the reasons we like doing this stuff!) But it can take time away from "road-testing" the existing functionality out in the real world. My sense is that you'll get more out of focusing on that for now.

> Perhaps this specific question about code organization/repetition is actually in scope for the review process?

Yes. Let's do the following:
- [ ] Please reply to my comment on the submission, saying you have come to a decision on most of the questions except the one you posted in the forum
- [ ] In that same comment, link to the question in the forum, and say you would like reviewers to take that into consideration as they do the review
- [ ] If you make changes based on feedback in the forum, you can publish a [release candidate](https://peps.python.org/pep-0440/#pre-releases) so that reviewers have access to these changes during the review **without** you releasing a new full version before we finish
