+++ CHANGELOG.rst
@@ -7,7 +7,7 @@
   formatting algorithm code.
 * Remove ``float``-centered language fom documentation.
 * Include a discussion in the documentation about ``Decimal`` versus
-  ``float`` considerations that my be important for users.
+  ``float`` considerations that may be important for users.
 
 0.21.0 (2023-07-22)
 -------------------
