### 2023-07-26 ea2a214
changelog removing arithmetic for SciNum
