### 2023-07-26 73b03fb
Don't include round_mode in README doctest for val/unc formatting since it has no effect (except raising a warning if it is not SIG_FIG)
