### 2023-07-26 e7e7315
minor docstring cleanup
