+++ src/sciform/format_options.py
@@ -509,9 +509,9 @@ def global_add_c_prefix():
 
 def global_add_small_si_prefixes():
     """
-    Include {-2: 'c', -1: 'd', +1: 'da', +2: 'h'} exponnet value to
-    prefix subsitutions. Note, if any of these exponent values are
-    mapped then that mapping will NOT be overwritten. To modify existing
+    Include ``{-2: 'c', -1: 'd', +1: 'da', +2: 'h'}`` to prefix
+    substitutions. Note, if any of these exponent values are mapped,
+    then that mapping will NOT be overwritten. To modify existing
     mappings either first use :func:`global_reset_si_prefixes` or use
     :func:`set_global_defaults`.
     """
