### 2023-07-27 18a3406
options documentation improvements
