+++ docs/source/project.rst
@@ -10,8 +10,9 @@ Changelog
 Roadmap
 =======
 
-* Support for parsing strings back into ``SciNum`` of ``SciNumUnc``
-  objects
+* Support for parsing strings into ``SciNum`` of ``SciNumUnc`` objects
+* Refactor API so that users construct and pass ``FormatOptions``-like
+  classes rather than passing ``kwargs`` directly.
 * Consider support for scientific units
 * Consider support for helper utility features such as formatting of
   Sequences or arrays and arithmetic of ``SciNum`` objects.
