+++ CHANGELOG.rst
@@ -10,25 +10,30 @@
   ``float`` and configuring ``SciNum`` to behave as a subclass of
   ``Decimal`` would require added complexity.
   Second, A decision has been made to keep the ``sciform`` module
-  focussed simply on formatting individual numbers or pairs of numbers
+  focussed solely on formatting individual numbers or pairs of numbers
   for early releases.
   Convenience functionality outside of this narrow scope will be
   considered at a later time.
 * Favor ``Decimal`` methods over ``float`` methods in internal
   formatting algorithm code.
-* Remove ``float``-centered language fom documentation.
-* Include a discussion in the documentation about ``Decimal`` versus
-  ``float`` considerations that may be important for users.
-* Add "under construction" message to README.
+* Documentation
+
+   * Remove ``float``-based language fom documentation.
+   * Include a discussion in the documentation about ``Decimal`` versus
+     ``float`` considerations that may be important for users.
+   * Various minor revisions and edits. Notably a typo in the version
+     ``0.21.0`` changelog entry that reversed the meaning of a sentence
+     was corrected.
+   * Add "under construction" message to README.
 
 0.21.0 (2023-07-22)
 -------------------
 
 * Use ``Decimal`` under the hood for numerical formatting instead of
-  ``float``. ``Decimal`` support higher precision than ``float`` and
-  more reliable rounding behavior.
+  ``float``. ``Decimal`` instances support higher precision than
+  ``float`` and more reliable rounding behavior.
 * Update particle data group uncertainty rounding unit tests since edge
-  cases are not handled property as a result of adopting ``Decimal``.
+  cases are now handled property as a result of adopting ``Decimal``.
 * Minor cleanup of ``sfloat`` arithemetic functions.
 
 0.20.1 (2023-06-24)
@@ -55,7 +60,7 @@
 * When an exponent translation mode is used in combination with Latex
   mode, the translated exponent will now be wrapped in a Latex text
   mode: e.g. ``\text{Mi}``.
-* Link to test cases on examples page
+* Link to test cases on examples page.
 
 0.19.0 (2023-06-22)
 -------------------
