+++ docs/source/project.rst
@@ -10,9 +10,9 @@ Changelog
 Roadmap
 =======
 
-* Support for parsing strings into ``SciNum`` of ``SciNumUnc`` objects
+* Support for parsing strings into ``SciNum`` of ``SciNumUnc`` objects.
 * Refactor API so that users construct and pass ``FormatOptions``-like
   classes rather than passing ``kwargs`` directly.
-* Consider support for scientific units
+* Consider support for scientific units.
 * Consider support for helper utility features such as formatting of
   sequences or arrays and arithmetic of ``SciNum`` objects.
