### 2023-07-27 a576a1d
usage cleanup. Especially improvement of float/decimal section.
