+++ src/sciform/formatter.py
@@ -36,19 +36,21 @@ class Formatter:
     The following checks are performed when creating a
     :class:`Formatter` object:
 
-    * precision >= 1 for significant figure rounding mode
-    * exp must be consistent with the exponent mode:
+    * ``precision`` >= 1 for significant figure rounding mode
+    * ``exp`` must be consistent with the exponent mode:
 
-      * exp must be 0 for fixed point and percent modes
-      * exp must be a multiple of 3 for engineering and shifted
+      * ``exp`` must be 0 for fixed point and percent modes
+      * ``exp`` must be a multiple of 3 for engineering and shifted
         engineering modes
-      * exp must be a multiple of 10 for binary iec mode
+      * ``exp`` must be a multiple of 10 for binary iec mode
 
-    * upper_separator must be different from decimal_separator
-    * decimal_separator may only be GroupingSeparator.POINT or
-      GroupingSeparator.COMMA
-    * lower_separator may only be GroupingSeparator.NONE,
-      GroupingSeparator.SPACE, or GroupingSeparator.UNDERSCORE.
+    * ``upper_separator`` may be any :class:`GroupingSeparator` but must
+      be different from ``decimal_separator``
+    * ``decimal_separator`` may only be :class:`GroupingSeparator.POINT`
+      or :class:`GroupingSeparator.COMMA`
+    * ``lower_separator`` may only be :class:`GroupingSeparator.NONE`,
+      :class:`GroupingSeparator.SPACE`, or
+      :class:`GroupingSeparator.UNDERSCORE`
 
     :param fill_mode: :class:`FillMode` indicating whether
       to fill with zeros or spaces.
@@ -57,7 +59,7 @@ class Formatter:
     :param top_dig_place: Positive ``int`` indicating the digits place
       to which the string will be left padded before the sign symbol. 0
       corresponds to the ones place, 1 corresponds to the tens place
-      etc. e.g. top_dig_place=4 will convert ``12`` into ``00012``.
+      etc. E.g. ``top_dig_place=4`` will convert ``12`` into ``00012``.
     :param upper_separator: :class:`GroupingSeparator` indicating the
       character to be used to group digits above the decimal symbol.
     :param decimal_separator: :class:`GroupingSeparator` indicating
@@ -68,78 +70,79 @@ class Formatter:
     :param lower_separator: :class:`GroupingSeparator` indicating the
       character to be used to group digits below the decimal symbol.
       :class:`GroupingSeparator.NONE`, :class:`GroupingSeparator.SPACE`,
-      :class:`GroupingSeparator.UNDERSCORE`.
+      or :class:`GroupingSeparator.UNDERSCORE`.
     :param round_mode: :class:`RoundMode` indicating whether to round
       the number based on significant figures or digits past the
       decimal point
-    :param precision: ``int`` or :class:`AutoPrec` sentinel indicating
+    :param precision: :class:`int` or :class:`AutoPrec` sentinel indicating
       how many significant figures or digits past the decimal point to
       include for rounding. Must be >= 1 for significant figure
       rounding. May be any integer for digits-past-the-decimal rounding.
     :param exp_mode: :class:`ExpMode` indicating the formatting
       mode to be used.
-    :param exp: ``int`` or :class:`AutoExp` indicating the value which
+    :param exp: :class:`int` or :class:`AutoExp` indicating the value which
       should be used for the exponent. This parameter is ignored for the
       fixed point exponent mode. For engineering, engineering shifted,
       and binary iec modes, if this parameter is not consistent with the
       rules of that mode (e.g. if it is not a multiple of 3), then the
       exponent is rounded down to the nearest conforming value and a
       warning is printed.
[…]
