### 2023-07-27 b32bb82
add requirements for docs build and readthedocs.yaml
