+++ docs/requirements.txt
@@ -1,1 +1,3 @@
 -e .
+sphinx==6.2.1
+sphinx-rtd-theme==1.2.2
