### 2023-07-28 d899cc7
float -> number in pyproject.toml description
