### 2023-07-29 9f6ad2f
major FormatOptions code refactor
- User does all configurations through FormatOptions class. (Formatter, global configs)
- FormatOptions is "rendered" into RenderedFormatOptions by patching unspecified options with options from global defaults.
- Documentation needs to be refactored
- test scripts are passing
- Doc tests are failing
