### 2023-07-29 49de6a9
fix AutoExp bug in FormatOptions
- related to plan to change Sentinels from AutoExp class to AutoExp class instances
