### 2023-07-29 b85b375
Make it so FormatOptions template cannot be RenderedFormatOptions again
