### 2023-07-29 3679e6a
Back to None sentinel value
