+++ src/sciform/format_options.py
@@ -38,23 +38,6 @@ class RenderedFormatOptions:
     unc_pm_whitespace: bool
 
 
-class Free:
-    """
-    Sentinel class indicating that a particular option in FormatOptions
-    is unpopulated. `Free` instances will be replaced with global
-    defaults upon rendering of FormatOptions into RenderedFormatOptions.
-    """
-    pass
-
-
-def is_free(obj) -> bool:
-    return isinstance(obj, Free)
-
-
-def is_not_free(obj) -> bool:
-    return not is_free(obj)
-
-
 ExpReplaceDict = dict[int, Union[str, None]]
 
 
@@ -63,32 +46,32 @@ class FormatOptions:
             self,
             *,
             template: Union['FormatOptions'] = None,
-            exp_mode: Union[ExpMode, Free] = Free(),
-            exp: Union[int, type(AutoExp), Free] = Free(),
-            percent: Union[bool, Free] = Free(),
-            round_mode: Union[RoundMode, Free] = Free(),
-            precision: Union[int, type(AutoPrec), Free] = Free(),
-            upper_separator: Union[UpperGroupingSeparators, Free] = Free(),
-            decimal_separator: Union[DecimalGroupingSeparators, Free] = Free(),
-            lower_separator: Union[LowerGroupingSeparators, Free] = Free(),
-            sign_mode: Union[SignMode, Free] = Free(),
-            fill_mode: Union[FillMode, Free] = Free(),
-            top_dig_place: Union[int, Free] = Free(),
-            prefix_exp: Union[bool, Free] = Free(),
-            parts_per_exp: Union[bool, Free] = Free(),
-            extra_si_prefixes: Union[ExpReplaceDict, Free] = Free(),
-            extra_iec_prefixes: Union[ExpReplaceDict, Free] = Free(),
-            extra_parts_per_forms: Union[ExpReplaceDict, Free] = Free(),
-            capitalize: Union[bool, Free] = Free(),
-            superscript_exp: Union[bool, Free] = Free(),
-            latex: Union[bool, Free] = Free(),
-            nan_inf_exp: Union[bool, Free] = Free(),
-            bracket_unc: Union[bool, Free] = Free(),
-            pdg_sig_figs: Union[bool, Free] = Free(),
-            val_unc_match_widths: Union[bool, Free] = Free(),
-            bracket_unc_remove_seps: Union[bool, Free] = Free(),
-            unicode_pm: Union[bool, Free] = Free(),
-            unc_pm_whitespace: Union[bool, Free] = Free(),
+            exp_mode: ExpMode = None,
+            exp: Union[int, type(AutoExp)] = None,
+            percent: bool = None,
+            round_mode: RoundMode = None,
+            precision: Union[int, type(AutoPrec)] = None,
+            upper_separator: UpperGroupingSeparators = None,
+            decimal_separator: DecimalGroupingSeparators = None,
+            lower_separator: LowerGroupingSeparators = None,
+            sign_mode: SignMode = None,
+            fill_mode: FillMode = None,
+            top_dig_place: int = None,
+            prefix_exp: bool = None,
+            parts_per_exp: bool = None,
+            extra_si_prefixes: ExpReplaceDict = None,
+            extra_iec_prefixes: ExpReplaceDict = None,
+            extra_parts_per_forms: ExpReplaceDict = None,
+            capitalize: bool = None,
+            superscript_exp: bool = None,
+            latex: bool = None,
+            nan_inf_exp: bool = None,
+            bracket_unc: bool = None,
+            pdg_sig_figs: bool = None,
+            val_unc_match_widths: bool = None,
+            bracket_unc_remove_seps: bool = None,
+            unicode_pm: bool = None,
[…]
