+++ src/sciform/fsml.py
@@ -1,6 +1,6 @@
 import re
 
-from sciform.format_options import FormatOptions, Free
+from sciform.format_options import FormatOptions
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
                            ExpMode)
 
@@ -22,23 +22,23 @@ pattern = re.compile(r'''^
 
 fill_mode_mapping = {' ': FillMode.SPACE,
                      '0': FillMode.ZERO,
-                     None: Free()}
+                     None: None}
 
 sign_mode_mapping = {'-': SignMode.NEGATIVE,
                      '+': SignMode.ALWAYS,
                      ' ': SignMode.SPACE,
-                     None: Free()}
+                     None: None}
 
 separator_mapping = {'n': GroupingSeparator.NONE,
                      ',': GroupingSeparator.COMMA,
                      '.': GroupingSeparator.POINT,
                      's': GroupingSeparator.SPACE,
                      '_': GroupingSeparator.UNDERSCORE,
-                     None: Free()}
+                     None: None}
 
 round_mode_mapping = {'!': RoundMode.SIG_FIG,
                       '.': RoundMode.PREC,
-                      None: Free()}
+                      None: None}
 
 
 def format_options_from_fmt_spec(fmt_spec: str) -> 'FormatOptions':
@@ -50,13 +50,13 @@ def format_options_from_fmt_spec(fmt_spec: str) -> 'FormatOptions':
     if fill_mode_flag is not None:
         fill_mode = fill_mode_mapping[fill_mode_flag]
     else:
-        fill_mode = Free()
+        fill_mode = None
 
     sign_mode_flag = match.group('sign_mode')
     if sign_mode_flag is not None:
         sign_mode = sign_mode_mapping[sign_mode_flag]
     else:
-        sign_mode = Free()
+        sign_mode = None
 
     alternate_mode = match.group('alternate_mode')
     alternate_mode = alternate_mode is not None
@@ -66,8 +66,8 @@ def format_options_from_fmt_spec(fmt_spec: str) -> 'FormatOptions':
         top_dig_place = int(top_dig_place)
         val_unc_match_widths = True
     else:
-        top_dig_place = Free()
-        val_unc_match_widths = Free()
+        top_dig_place = None
+        val_unc_match_widths = None
 
     upper_separator_flag = match.group('upper_separator')
     upper_separator = separator_mapping[upper_separator_flag]
@@ -85,9 +85,9 @@ def format_options_from_fmt_spec(fmt_spec: str) -> 'FormatOptions':
     if precision is not None:
         precision = int(precision)
     else:
-        precision = Free()
+        precision = None
 
-    percent = Free()
+    percent = None
     exp_mode = match.group('exp_mode')
     if exp_mode is not None:
         capitalize = exp_mode.isupper()
@@ -108,26 +108,26 @@ def format_options_from_fmt_spec(fmt_spec: str) -> 'FormatOptions':
             else:
                 exp_mode = ExpMode.BINARY
     else:
-        capitalize = Free()
-        exp_mode = Free()
+        capitalize = None
+        exp_mode = None
 
     exp = match.group('exp')
     if exp is not None:
         exp = int(exp)
     else:
-        exp = Free()
+        exp = None
 
     prefix_exp = match.group('prefix_mode')
     if prefix_exp is not None:
         prefix_exp = True
     else:
-        prefix_exp = Free()
+        prefix_exp = None
 
     bracket_unc = match.group('bracket_unc')
     if bracket_unc is not None:
         bracket_unc = True
     else:
-        bracket_unc = Free()
+        bracket_unc = None
 
     return FormatOptions(
         fill_mode=fill_mode,
