### 2023-07-29 8e71772
Update FormatOptions and Formatter docstrings
