+++ src/sciform/format_options.py
@@ -42,6 +42,139 @@ ExpReplaceDict = dict[int, Union[str, None]]
 
 
 class FormatOptions:
+    """
+    :class:`FormatOptions` stores all the configuration options used to
+    format numbers and number/uncertainty pairs. See
+    :ref:`formatting_options` for more details on the available options.
+    :class:`FormatOptions` are used to create :class:`Formatter`
+    instances and to modify the global default configuration.
+
+    It is not necessary to provide input for all options. There are two
+    mechanisms for filling of any unsupplied options. First, the user
+    can pass in another :class:`FormatOptions` instance as a
+    ``template``. In this case any populated options for the
+    ``template`` will be used to populate corresponding unpopulated
+    options for the new :class:`FormatOptions`. Second, at format time
+    any remaining unfilled options will be populated with the global
+    default options. See :ref:`global_config` for details about how to
+    view and modify the global default options.
+
+    The following checks are performed when creating a new
+    :class:`FormatOptions` object:
+
+    * ``precision`` >= 1 for significant figure rounding mode
+    * ``exp`` must be consistent with the exponent mode:
+
+      * ``exp`` must be 0 for fixed point and percent modes
+      * ``exp`` must be a multiple of 3 for engineering and shifted
+        engineering modes
+      * ``exp`` must be a multiple of 10 for binary iec mode
+
+    * ``upper_separator`` may be any :class:`GroupingSeparator` but must
+      be different from ``decimal_separator``
+    * ``decimal_separator`` may only be :class:`GroupingSeparator.POINT`
+      or :class:`GroupingSeparator.COMMA`
+    * ``lower_separator`` may only be :class:`GroupingSeparator.NONE`,
+      :class:`GroupingSeparator.SPACE`, or
+      :class:`GroupingSeparator.UNDERSCORE`
+    * Only one of ``prefix_exp`` and ``parts_per_exp`` may be selected.
+
+    :param template: :class:`FormatOptions` instance to use to populate
+      unfilled options.
+    :param exp_mode: :class:`ExpMode` indicating the formatting
+      mode to be used.
+    :param exp: :class:`int` or :class:`AutoExp` indicating the value which
+      should be used for the exponent. This parameter is ignored for the
+      fixed point exponent mode. For engineering, engineering shifted,
+      and binary iec modes, if this parameter is not consistent with the
+      rules of that mode (e.g. if it is not a multiple of 3), then the
+      exponent is rounded down to the nearest conforming value and a
+      warning is printed.
+    :param percent: :class:`bool` indicating whether the number should be
+      formatted as a percentage or not. Only valid for fixed point
+      exponent mode. When ``True``, the number is multipled by 100 and
+      a ``%`` symbol is appended to the end of the string after
+      formatting.
+    :param round_mode: :class:`RoundMode` indicating whether to round
+      the number based on significant figures or digits past the
+      decimal point
+    :param precision: :class:`int` or :class:`AutoPrec` sentinel indicating
+      how many significant figures or digits past the decimal point to
+      include for rounding. Must be >= 1 for significant figure
+      rounding. May be any integer for digits-past-the-decimal rounding.
+    :param upper_separator: :class:`GroupingSeparator` indicating the
+      character to be used to group digits above the decimal symbol.
+    :param decimal_separator: :class:`GroupingSeparator` indicating
+      the character to be used as the decimal symbol.
+      :class:`GroupingSeparator.POINT` or
[…]
