+++ src/sciform/formatter.py
@@ -6,140 +6,32 @@ from sciform.format_utils import Number
 
 class Formatter:
     """
-    Formatter object used to convert numbers and pairs of numbers into
-    formatted strings. See :ref:`formatting_options` for more details on
-    the available options. Any options which are not provided by the
-    user will be filled with the corresponding values from the global
-    default configuration.
+    :class:`Formatter` is used to convert numbers and pairs of numbers
+    into formatted strings. Formatting options are configured using
+    :class:`FormatOptions`. Any unpopulated format options will be
+    populated from the global default options at format time. See
+    :ref:`formatting_options` for more details on the available options.
 
-    >>> from sciform import Formatter, ExpMode, RoundMode
-    >>> sform = Formatter(exp_mode=ExpMode.ENGINEERING,
-    ...                   round_mode=RoundMode.SIG_FIG,
-    ...                   precision=4)
+    >>> from sciform import FormatOptions, Formatter, ExpMode, RoundMode
+    >>> sform = Formatter(FormatOptions(
+    ...             exp_mode=ExpMode.ENGINEERING,
+    ...             round_mode=RoundMode.SIG_FIG,
+    ...             precision=4))
     >>> print(sform(12345.678))
     12.35e+03
 
     The Formatter can be called with two aguments for value/uncertainty
     formatting
 
-    >>> sform = Formatter(exp_mode=ExpMode.ENGINEERING,
-    ...                   round_mode=RoundMode.SIG_FIG,
-    ...                   precision=2)
+    >>> sform = Formatter(FormatOptions(
+    ...             exp_mode=ExpMode.ENGINEERING,
+    ...             round_mode=RoundMode.SIG_FIG,
+    ...             precision=2))
     >>> print(sform(12345.678, 3.4))
     (12.3457 +/- 0.0034)e+03
 
-    The following checks are performed when creating a
-    :class:`Formatter` object:
-
-    * ``precision`` >= 1 for significant figure rounding mode
-    * ``exp`` must be consistent with the exponent mode:
-
-      * ``exp`` must be 0 for fixed point and percent modes
-      * ``exp`` must be a multiple of 3 for engineering and shifted
-        engineering modes
-      * ``exp`` must be a multiple of 10 for binary iec mode
-
-    * ``upper_separator`` may be any :class:`GroupingSeparator` but must
-      be different from ``decimal_separator``
-    * ``decimal_separator`` may only be :class:`GroupingSeparator.POINT`
-      or :class:`GroupingSeparator.COMMA`
-    * ``lower_separator`` may only be :class:`GroupingSeparator.NONE`,
-      :class:`GroupingSeparator.SPACE`, or
-      :class:`GroupingSeparator.UNDERSCORE`
-
-    :param fill_mode: :class:`FillMode` indicating whether
-      to fill with zeros or spaces.
-    :param sign_mode: :class:`SignMode` indicating sign
-      symbol behavior.
-    :param top_dig_place: Positive ``int`` indicating the digits place
-      to which the string will be left padded before the sign symbol. 0
-      corresponds to the ones place, 1 corresponds to the tens place
-      etc. E.g. ``top_dig_place=4`` will convert ``12`` into ``00012``.
-    :param upper_separator: :class:`GroupingSeparator` indicating the
-      character to be used to group digits above the decimal symbol.
-    :param decimal_separator: :class:`GroupingSeparator` indicating
-      the character to be used as the decimal symbol.
-      :class:`GroupingSeparator.POINT` or
-      :class:`GroupingSeparator.COMMA`. Note that ``decimal_separator``
-      cannot be the same as ``upper_separator``
-    :param lower_separator: :class:`GroupingSeparator` indicating the
-      character to be used to group digits below the decimal symbol.
[…]
