### 2023-07-29 75b8c79
Update readme with new FormatOptions structure
