### 2023-07-29 fd67412
If no options are passed into Formatter then global defaults are used
