### 2023-07-29 a624e21
format_options update doctests to new usage
