+++ docs/source/options.rst
@@ -33,8 +33,9 @@ Fixed Point
 Fixed point notation is standard notation in which a number is displayed
 directly with no extra exponent.
 
+>>> from sciform import FormatOptions as Fo
 >>> from sciform import Formatter, ExpMode, RoundMode
->>> sform = Formatter(exp_mode=ExpMode.FIXEDPOINT)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.FIXEDPOINT))
 >>> print(sform(123.456))
 123.456
 
@@ -47,7 +48,7 @@ Scientific notation is used to display base-10 decimal numbers.
 In scientific notation, the exponent is uniquely chosen so that the
 mantissa ``m`` satisfies ``1 <= m < 10``.
 
->>> sform = Formatter(exp_mode=ExpMode.SCIENTIFIC)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.SCIENTIFIC))
 >>> print(sform(123.456))
 1.23456e+02
 
@@ -76,7 +77,7 @@ Here it may be more convenient to display the number with an exponent of
 to an exponent of 8 which would be chosen for standard scientific
 notation.
 
->>> sform = Formatter(exp_mode=ExpMode.ENGINEERING)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.ENGINEERING))
 >>> print(sform(123.456))
 123.456e+00
 
@@ -89,7 +90,7 @@ Shifted engineering notation is the same as engineering notation except
 the exponent is chosen so that the mantissa ``m`` satisfies
 ``0.1 <= m < 100``.
 
->>> sform = Formatter(exp_mode=ExpMode.ENGINEERING_SHIFTED)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.ENGINEERING_SHIFTED))
 >>> print(sform(123.456))
 0.123456e+03
 
@@ -101,7 +102,7 @@ Binary
 Binary formatting can be chosen to display a number in scientific
 notation in base-2.
 
->>> sform = Formatter(exp_mode=ExpMode.BINARY)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.BINARY))
 >>> print(sform(256))
 1b+08
 
@@ -122,7 +123,7 @@ The mantissa ``m`` satisfies ``1 <= m < 1024``.
 Binary formatting can be chosen to display a number in scientific
 notation in base-2.
 
->>> sform = Formatter(exp_mode=ExpMode.BINARY_IEC)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.BINARY_IEC))
 >>> print(sform(2048))
 2b+10
 
@@ -135,8 +136,8 @@ If a fixed exponent is requested for engineering, shifted engineering,
 or binary IEC mode, then the requested exponent is rounded down to the
 nearest multiple of 3 or 10.
 
->>> sform = Formatter(exp_mode=ExpMode.SCIENTIFIC,
-...                   exp=3)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.SCIENTIFIC,
+...                      exp=3))
 >>> print(sform(123.456))
 0.123456e+03
 
@@ -160,18 +161,18 @@ Furthermore, it is possible to customize :class:`Formatter`
 objects or the global configuration settings to map additional
 translations, in addition to those provided by default.
 
->>> sform = Formatter(exp_mode=ExpMode.ENGINEERING,
-...                   prefix_exp=True)
+>>> sform = Formatter(Fo(exp_mode=ExpMode.ENGINEERING,
+...                      prefix_exp=True))
 >>> print(sform(4242.13))
 4.24213 k
[…]
