### 2023-07-29 1d317ef
Update exp replacement doc tests
