### 2023-07-29 7de4356
format options usage documentation
