+++ docs/source/usage.rst
@@ -45,16 +45,17 @@ a corresponding formatted string.
 Note that global default options are used to populate unfilled options
 at formatting time.
 
+>>> from sciform import FormatOptions as Fo
 >>> from sciform import Formatter, RoundMode, GroupingSeparator, ExpMode
->>> sform = Formatter(round_mode=RoundMode.PREC,
-...                   precision=6,
-...                   upper_separator=GroupingSeparator.SPACE,
-...                   lower_separator=GroupingSeparator.SPACE)
+>>> sform = Formatter(Fo(round_mode=RoundMode.PREC,
+...                      precision=6,
+...                      upper_separator=GroupingSeparator.SPACE,
+...                      lower_separator=GroupingSeparator.SPACE))
 >>> print(sform(51413.14159265359))
 51 413.141 593
->>> sform = Formatter(round_mode=RoundMode.SIG_FIG,
-...                   precision=4,
-...                   exp_mode=ExpMode.ENGINEERING)
+>>> sform = Formatter(Fo(round_mode=RoundMode.SIG_FIG,
+...                      precision=4,
+...                      exp_mode=ExpMode.ENGINEERING))
 >>> print(sform(123456.78))
 123.5e+03
 
@@ -83,16 +84,16 @@ value/uncertainty strings.
 :mod:`sciform` attempts to follow
 `BIPM <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6>`_
 or `NIST <https://www.nist.gov/pml/nist-technical-note-1297>`_
-recomendations for conventions when possible.
+recommendations for conventions when possible.
 
-Value/uncertainty pairs can be formatted either by passing two values
+Value/uncertainty pairs can be formatted either by passing two numbers
 into a :class:`Formatter`, and configuring the :class:`Formatter` using
 :ref:`formatting_options` and :ref:`val_unc_formatting_options`, or by
 using the :class:`SciNumUnc` object.
 
 >>> val = 84.3
 >>> unc = 0.2
->>> sform = Formatter(precision=2)
+>>> sform = Formatter(Fo(precision=2))
 >>> print(sform(val, unc))
 84.30 +/- 0.20
 
@@ -114,7 +115,7 @@ algorithm:
 #. Rounding is always performed using significant figure rounding
    applied to the uncertainty. If a ``precision`` is supplied then the
    uncertainty is rounded to significant figures consistent with the
-   supplied ``precision``. Otherwise the uncertainty is left unrounded.
+   supplied ``precision``. Otherwise the uncertainty is left un-rounded.
 #. The value is rounded to the digit corresponding to the least
    significant digit of the rounded uncertainty.
 #. The value for the exponent is resolved by applying the
@@ -144,28 +145,28 @@ package default settings):
 
 >>> from sciform import print_global_defaults
 >>> print_global_defaults()
-{'fill_mode': <FillMode.SPACE: 'space'>,
- 'sign_mode': <SignMode.NEGATIVE: 'negative'>,
- 'top_dig_place': 0,
+{'exp_mode': <ExpMode.FIXEDPOINT: 'fixed_point'>,
+ 'exp': <class 'sciform.modes.AutoExp'>,
+ 'percent': False,
[…]
