### 2023-07-30 3ff9f37
global configuration documentation
