+++ CHANGELOG.rst
@@ -6,6 +6,11 @@
   functions.
   ``Formatter`` and global configuration functions no longer accept bare
   keyword arguments to indicate formatting options.
+* **[BREAKING]** ``Formatter`` now resolves unspecified format options
+  from the global defaults at format time instead of initialization
+  time.
+  This is more similar to the previous behavior for ``SciNum`` and
+  ``SciNumUnc``.
 * Change ``pyproject.toml`` description
 
 
