### 2023-07-30 d3ed04d
Revert "exp_mode percent documentation"
This reverts commit 5b3295877e49a958fd0a9a06c72214ffe5395fd5.
