+++ tests/test_val_unc_formatter.py
@@ -43,10 +43,10 @@ class TestFormatting(unittest.TestCase):
     def test_percent(self):
         cases_list = [
             ((0.123_456_78, 0.000_002_55), [
-                (FormatOptions(percent=True,
+                (FormatOptions(exp_mode=ExpMode.PERCENT,
                                lower_separator=GroupingSeparator.UNDERSCORE),
                  '(12.345_678 +/- 0.000_255)%'),
-                (FormatOptions(percent=True,
+                (FormatOptions(exp_mode=ExpMode.PERCENT,
                                bracket_unc=True,
                                lower_separator=GroupingSeparator.UNDERSCORE),
                  '(12.345_678(255))%')
@@ -131,11 +131,11 @@ class TestFormatting(unittest.TestCase):
             ]),
             ((0.123_456_78, 0.000_002_55), [
                 (FormatOptions(lower_separator=GroupingSeparator.UNDERSCORE,
-                               percent=True,
+                               exp_mode=ExpMode.PERCENT,
                                latex=True),
                  r'\left(12.345\_678 \pm 0.000\_255\right)\%'),
                 (FormatOptions(lower_separator=GroupingSeparator.UNDERSCORE,
-                               percent=True,
+                               exp_mode=ExpMode.PERCENT,
                                bracket_unc=True,
                                latex=True),
                  r'\left(12.345\_678\left(255\right)\right)\%')
