### 2023-07-30 2e178bf
Some minor code changes to cleanup percent code
