+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.23.0"
+__version__ = "0.24.0"
 
 from sciform.format_options import FormatOptions
 from sciform.formatter import Formatter
