### 2023-07-31 022eb62
make format_options a dataclass and automate replace and render
