+++ src/sciform/format_options.py
@@ -1,5 +1,5 @@
 from typing import Union, get_args
-from dataclasses import dataclass, asdict
+from dataclasses import dataclass, asdict, InitVar, replace
 from pprint import pprint
 
 from sciform.modes import (FillMode, SignMode, GroupingSeparator,
@@ -40,6 +40,7 @@ class RenderedFormatOptions:
 ExpReplaceDict = dict[int, Union[str, None]]
 
 
+@dataclass(frozen=True)
 class FormatOptions:
     # TODO: __repr__
     """
@@ -170,196 +171,130 @@ class FormatOptions:
     :param add_ppth_form: :class:`bool` (default ``False``) if ``True`` adds
       ``{-3: 'ppth'}`` to ``extra_parts_per_forms``.
     """
-    def __init__(
-            self,
-            *,
-            template: Union['FormatOptions'] = None,
-            exp_mode: ExpMode = None,
-            exp: Union[int, type(AutoExp)] = None,
-            round_mode: RoundMode = None,
-            precision: Union[int, type(AutoPrec)] = None,
-            upper_separator: UpperGroupingSeparators = None,
-            decimal_separator: DecimalGroupingSeparators = None,
-            lower_separator: LowerGroupingSeparators = None,
-            sign_mode: SignMode = None,
-            fill_mode: FillMode = None,
-            top_dig_place: int = None,
-            prefix_exp: bool = None,
-            parts_per_exp: bool = None,
-            extra_si_prefixes: ExpReplaceDict = None,
-            extra_iec_prefixes: ExpReplaceDict = None,
-            extra_parts_per_forms: ExpReplaceDict = None,
-            capitalize: bool = None,
-            superscript_exp: bool = None,
-            latex: bool = None,
-            nan_inf_exp: bool = None,
-            bracket_unc: bool = None,
-            pdg_sig_figs: bool = None,
-            val_unc_match_widths: bool = None,
-            bracket_unc_remove_seps: bool = None,
-            unicode_pm: bool = None,
-            unc_pm_whitespace: bool = None,
-            add_c_prefix: bool = False,
-            add_small_si_prefixes: bool = False,
-            add_ppth_form: bool = False
-    ):
-        if round_mode is RoundMode.SIG_FIG:
-            if isinstance(precision, int):
-                if precision < 1:
-                    raise ValueError(f'Precision must be >= 1 for sig fig '
-                                     f'rounding, not {precision}.')
+    exp_mode: ExpMode = None
+    exp: Union[int, type(AutoExp)] = None
+    round_mode: RoundMode = None
+    precision: Union[int, type(AutoPrec)] = None
+    upper_separator: UpperGroupingSeparators = None
+    decimal_separator: DecimalGroupingSeparators = None
+    lower_separator: LowerGroupingSeparators = None
+    sign_mode: SignMode = None
+    fill_mode: FillMode = None
+    top_dig_place: int = None
+    prefix_exp: bool = None
+    parts_per_exp: bool = None
+    extra_si_prefixes: dict[int, str] = None
+    extra_iec_prefixes: dict[int, str] = None
+    extra_parts_per_forms: dict[int, str] = None
+    capitalize: bool = None
+    superscript_exp: bool = None
+    latex: bool = None
+    nan_inf_exp: bool = None
+    bracket_unc: bool = None
+    pdg_sig_figs: bool = None
+    val_unc_match_widths: bool = None
+    bracket_unc_remove_seps: bool = None
+    unicode_pm: bool = None
+    unc_pm_whitespace: bool = None
 
-        if exp is not AutoExp and exp is not None:
[…]
