### 2023-07-31 43e81b7
render accepts default options, DEFAULT_PKG_OPTIONS -> PKG_DEFAULT_OPTIONS for consistency with GLOBAL_DEFAULT_OPTIONS
