### 2023-07-31 3f69534
FormatOptions remove template documentation
