+++ docs/source/usage.rst
@@ -23,17 +23,11 @@ See :class:`FormatOptions` for a complete list of
 keyword arguments used to construct :class:`FormatOptions` and see
 :ref:`formatting_options` for more details on the different options.
 
-When constructing a :class:`FormatOptions` object, it is not necessary
-to provide input for all options.
-There are two mechanisms for filling of any un-supplied options.
-First, during initialization, the user can pass in another
-:class:`FormatOptions` instance as a ``template``.
-In this case any populated options for the ``template`` will be used to
-populate corresponding unpopulated options for the new
-:class:`FormatOptions`.
-Second, at format time any remaining unfilled options will be populated
-with the global default options. See :ref:`global_config` for details
-about how to view and modify the global default options.
+It is not necessary to provide input for all options. At format time,
+any un-populated options will be populated with the
+corresponding options from the global default options.
+See :ref:`global_config` for details about how to view and modify the
+global default options.
 
 Formatter
 ---------
