### 2023-08-03 e085be3
replace and render use a shared implementation for merging dictionaries
