+++ docs/source/usage.rst
@@ -22,6 +22,11 @@ FormatOptions
 See :class:`FormatOptions` for a complete list of
 keyword arguments used to construct :class:`FormatOptions` and see
 :ref:`formatting_options` for more details on the different options.
+The :meth:`FormatOptions.replace` method can be used to construct new
+:class:`FormatOptions` instances by using one existing
+:class:`FormatOptions` instance as a base set of options and to use a
+second :class:`FormatOptions` instance to specify which options should
+be overwritten and with what values for the new instance.
 
 It is not necessary to provide input for all options. At format time,
 any un-populated options will be populated with the
