+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.25.0(2023-07-31)
+------------------
+
+* **[BREAKING]** ``template`` option remove from ``FormatOptions``
+  constructor.
+  New ``FormatOptions`` instances can be constructed using the
+  configuration of an existing ``FormatOptions`` instance by calling
+  the `replace()` method on the existing ``FormatOptions``.
+
 0.24.0 (2023-07-30)
 -------------------
 
