### 2023-08-03 3d22f5a
try to improve code readability
