### 2023-08-11 b1b508f
split get_pdg_round_digit into its own function
