+++ CHANGELOG.rst
@@ -1,7 +1,7 @@
 0.25.1 (2023-08-10)
 -------------------
 
-* Refactor ``get_pdg_sig_figs()`` into a dedicated function.
+* Refactor ``get_pdg_round_digit()`` into a dedicated function.
 
 0.25.0 (2023-08-02)
 ------------------
