+++ README.rst
@@ -57,7 +57,7 @@ strings according to the selected options.
 >>> from sciform import (FormatOptions, Formatter, RoundMode,
 ...                      GroupingSeparator, ExpMode)
 >>> sform = Formatter(FormatOptions(
-...             round_mode=RoundMode.PREC,
+...             round_mode=RoundMode.DIG_PLACE,
 ...             precision=6,
 ...             upper_separator=GroupingSeparator.SPACE,
 ...             lower_separator=GroupingSeparator.SPACE))
@@ -75,7 +75,7 @@ For brevity, the user may consider abbreviating ``FormatOptions`` as
 
 >>> from sciform import FormatOptions as Fo
 >>> from sciform import Formatter, RoundMode, GroupingSeparator, ExpMode
->>> sform = Formatter(Fo(round_mode=RoundMode.PREC,
+>>> sform = Formatter(Fo(round_mode=RoundMode.DIG_PLACE,
 ...                      precision=6,
 ...                      upper_separator=GroupingSeparator.SPACE,
 ...                      lower_separator=GroupingSeparator.SPACE))
