+++ src/sciform/format_utils.py
@@ -5,7 +5,7 @@ import re
 from copy import copy
 from decimal import Decimal
 
-from sciform.modes import ExpMode, RoundMode, SignMode, AutoExp, AutoPrec
+from sciform.modes import ExpMode, RoundMode, SignMode, AutoExp, AutoRound
 from sciform.prefix import (si_val_to_prefix_dict, iec_val_to_prefix_dict,
                             pp_val_to_prefix_dict)
 
@@ -166,10 +166,10 @@ def get_pdg_round_digit(num: Decimal):
 
 def get_round_digit(num: Decimal,
                     round_mode: RoundMode,
-                    precision: Union[int, type(AutoPrec)],
+                    precision: Union[int, type(AutoRound)],
                     pdg_sig_figs: bool = False) -> int:
     if round_mode is RoundMode.SIG_FIG:
-        if precision is AutoPrec:
+        if precision is AutoRound:
             if pdg_sig_figs:
                 round_digit = get_pdg_round_digit(num)
             else:
@@ -177,7 +177,7 @@ def get_round_digit(num: Decimal,
         else:
             round_digit = get_top_digit(num) - (precision - 1)
     elif round_mode is RoundMode.DIG_PLACE:
-        if precision is AutoPrec:
+        if precision is AutoRound:
             round_digit = get_bottom_digit(num)
         else:
             round_digit = -precision
