### 2023-08-13 f99825f
replace AutoPrec -> AutoRound in docs
