### 2023-08-13 c0ca9c1
rename exp to exp_val and AutoExp to AutoExpVal
