### 2023-08-14 c17cc06
replace RoundMode.DIG_PLACE with RoundMode.DEC_PLACE in doctests
