### 2023-08-14 09ede3d
disallow pdg_sig_figs=True without ndigits=None or ndigits=AutoRound
