+++ src/sciform/format_options.py
@@ -209,7 +209,14 @@ class FormatOptions:
                     raise ValueError(f'Precision must be >= 1 for sig fig '
                                      f'rounding, not {self.ndigits}.')
 
+        if (self.pdg_sig_figs and self.ndigits is not None
+                and self.ndigits is not AutoRound):
+            # TODO: test this
+            raise ValueError(f'pdg_sig_figs=True can only be used with '
+                             f'ndigits=AutoRound, not ndigits={self.ndigits}.')
+
         if self.exp_val is not AutoExpVal and self.exp_val is not None:
+            # TODO: Test these errors
             if (self.exp_mode is ExpMode.FIXEDPOINT
                     or self.exp_mode is ExpMode.PERCENT):
                 if self.exp_val != 0:
@@ -228,6 +235,7 @@ class FormatOptions:
                                      f'exp_val={self.exp_val}, for binary IEC '
                                      f'exponent mode.')
 
+        # TODO: Test all these separator errors
         if self.upper_separator is not None:
             if self.upper_separator not in get_args(UpperGroupingSeparators):
                 raise ValueError(f'upper_separator must be in '
@@ -255,6 +263,7 @@ class FormatOptions:
                 raise ValueError('Only one of prefix exponent and parts-per '
                                  'exponent modes may be selected.')
 
+        # TODO: Test that things do and don't get added appropriately
         if add_c_prefix:
             if self.extra_si_prefixes is None:
                 super().__setattr__('extra_si_prefixes', dict())
