### 2023-08-14 644bc0a
dig_place -> dec_place Enum value
