+++ docs/source/options.rst
@@ -21,7 +21,7 @@ formatting presents numbers in the form::
 
    num = mantissa * base**exp
 
-Where exp is an integer and ``base`` is typically 10 or 2.
+Where exp is an integer and ``base`` is 10 or 2.
 The different exponent modes control how ``mantissa``, ``base`` and
 ``exp`` are chosen for a given input number ``num``.
 
@@ -154,10 +154,6 @@ Fixed Exponent
 ==============
 
 The user can coerce the exponent for the formatting to a fixed value.
-The requested exponent is ignored in fixed point mode.
-If a fixed exponent is requested for engineering, shifted engineering,
-or binary IEC mode, then the requested exponent is rounded down to the
-nearest multiple of 3 or 10.
 
 >>> sform = Formatter(Fo(exp_mode=ExpMode.SCIENTIFIC,
 ...                      exp_val=3))
@@ -165,8 +161,20 @@ nearest multiple of 3 or 10.
 0.123456e+03
 
 To explicitly force :mod:`sciform` to automatically select the exponent
-then use the :class:`AutoExpVal` option by passing ``exp_val=AutoExpVal``.
+then use the :class:`AutoExpVal` option by passing
+``exp_val=AutoExpVal``.
 This is the default value in the global configuration.
+
+Note that the forced exponent must be consistent with the requested
+exponent mode.
+For fixed point and percent modes an explicit fixed exponent must equal
+0.
+For engineering and shifted engineering modes an explicit fixed exponent
+must be an integer multiple of 3.
+For binary IEC mode an explicit fixed exponent must be an integer
+multiple of 10.
+Because of this constrained behavior, it is recommended to only use a
+fixed exponent with the scientific or binary exponent modes.
 
 Exponent String Replacement
 ===========================
@@ -273,27 +281,31 @@ Rounding
 ========
 
 :mod:`sciform` provides two rounding strategies: rounding based on
-significant figures, and rounding based on digits past the decimal
-point or "precision".
+significant figures, and rounding based on decimal places.
 In both cases, the rounding applies to the mantissa determined after
 identifying the appropriate exponent for display based on the selected
 exponent mode.
 In some cases, the rounding results in a modification to the chosen
-exponent (e.g. presenting 9.99 with a precision of 1 in scientific
-exponent mode).
+exponent (e.g. when presenting ``9.99`` in scientific exponent mode with
+two digits past the decimal point we display  ``"9.99e+00"``, but with
+one digit past the decimal point we display ``"1.0e+01"``).
 This is taken into account before the final presentation.
 
-In both cases, if no explicit precision value or number of significant
-figures is supplied then all digits of the number are displayed.
-To explicitly force this behavior use the :class:`AutoRound` class by
-passing ``ndigits=AutoRound``.
+If the user does not specify the number of significant digits or the
+digits place to which to round, then the decimal numbers are displayed
+with full precision.
+To explicitly request this behavior, the user may use the
+:class:`AutoRound` sentinel by passing ``ndigits=AutoRound``.
 This is the default value in the global configuration.
 
-If the number to be formatted is passed in as a :class:`float` (either
-to a :class:`Formatter`, :class:`SciNum`, or :class:`SciNumUnc`),
-then, if no precision or number of significant figures is supplied, the
-decimal digits of the :class:`float` are truncated to the minimum number
-of digits necessary for round-tripping.
+Note that surprising behavior may be observed if using :class:`float`
+inputs.
[…]
