### 2023-08-14 16122b9
refactor GlobalDefaultsContext to use getter/setter
