### 2023-08-15 71c40fc
fsml variable renames and cleanup
