### 2023-08-15 6c3219a
refactor variables names and some program flow
