+++ src/sciform/format_utils.py
@@ -1,6 +1,5 @@
 from typing import Union
 from math import floor, log10, log2
-from warnings import warn
 import re
 from copy import copy
 from decimal import Decimal
@@ -32,14 +31,10 @@ def get_bottom_digit(num: Decimal) -> int:
         return exp
 
 
-def get_top_and_bottom_digit(num: Decimal) -> tuple[int, int]:
-    return get_top_digit(num), get_bottom_digit(num)
-
-
 def get_mantissa_exp_base(
         num: Decimal,
         exp_mode: ExpMode,
-        exp_val: Union[int, type(AutoExpVal)] = None
+        input_exp_val: Union[int, type(AutoExpVal)] = None
 ) -> (Decimal, int, int):
     if (exp_mode is ExpMode.BINARY
             or exp_mode is ExpMode.BINARY_IEC):
@@ -49,55 +44,62 @@ def get_mantissa_exp_base(
 
     if not num.is_finite():
         mantissa = num
-        if exp_val is AutoExpVal:
+        if input_exp_val is AutoExpVal:
             exp_val = 0
+        else:
+            exp_val = input_exp_val
     elif num == 0:
         mantissa = Decimal(0)
-        if exp_val is AutoExpVal:
+        if input_exp_val is AutoExpVal:
             exp_val = 0
+        else:
+            exp_val = input_exp_val
     else:
         if exp_mode is ExpMode.FIXEDPOINT or exp_mode is ExpMode.PERCENT:
-            if exp_val is not AutoExpVal:
-                if exp_val != 0:
-                    warn('Attempt to set non-zero exponent explicity in fixed '
-                         'point or percent exponent mode. Coercing exponent '
-                         'to 0.')
+            if input_exp_val is not AutoExpVal and input_exp_val != 0:
+                raise ValueError('Cannot set non-zero exponent in fixed point '
+                                 'or percent exponent mode.')
             exp_val = 0
         elif exp_mode is ExpMode.SCIENTIFIC:
-            if exp_val is AutoExpVal:
+            if input_exp_val is AutoExpVal:
                 exp_val = get_top_digit(num)
+            else:
+                exp_val = input_exp_val
         elif (exp_mode is ExpMode.ENGINEERING
               or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-            if exp_val is AutoExpVal:
+            if input_exp_val is AutoExpVal:
                 exp_val = get_top_digit(num)
                 if exp_mode is ExpMode.ENGINEERING:
                     exp_val = (exp_val // 3) * 3
                 else:
                     exp_val = ((exp_val + 1) // 3) * 3
             else:
-                if exp_val % 3 != 0:
-                    warn('Attempt to set exponent explicity to a non-integer '
-                         'multiple of 3 in engineering mode. Coercing to the '
-                         'next lower multiple of 3.')
-                    exp_val = (exp_val // 3) * 3
+                if input_exp_val % 3 != 0:
+                    raise ValueError(f'Exponent must be an integer multiple '
+                                     f'of 3 in engineering modes, not '
+                                     f'{input_exp_val}.')
+                exp_val = input_exp_val
         elif (exp_mode is ExpMode.BINARY
               or exp_mode is ExpMode.BINARY_IEC):
-            if exp_val is AutoExpVal:
+            if input_exp_val is AutoExpVal:
                 exp_val = get_top_digit(num, binary=True)
                 if exp_mode is ExpMode.BINARY_IEC:
                     exp_val = (exp_val // 10) * 10
             else:
-                if exp_mode is ExpMode.BINARY_IEC and exp_val % 10 != 0:
[…]
