+++ src/sciform/formatting.py
@@ -26,18 +26,18 @@ def format_non_inf(num: Decimal, options: RenderedFormatOptions) -> str:
     if options.nan_inf_exp:
         exp_mode = options.exp_mode
 
-        exp = options.exp_val
+        exp_val = options.exp_val
         if options.exp_val is AutoExpVal:
-            exp = 0
+            exp_val = 0
 
         if exp_mode is ExpMode.FIXEDPOINT:
             exp_str = ''
         elif (exp_mode is ExpMode.SCIENTIFIC
               or exp_mode is ExpMode.ENGINEERING
               or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-            exp_str = f'e+{exp:02d}'
+            exp_str = f'e+{exp_val:02d}'
         else:
-            exp_str = f'b+{exp:02d}'
+            exp_str = f'b+{exp_val:02d}'
 
         exp_str = convert_exp_str(exp_str,
                                   options.prefix_exp,
@@ -79,21 +79,23 @@ def format_num(num: Decimal, unrendered_options: FormatOptions) -> str:
         num *= 100
         num = num.normalize()
 
-    exp = options.exp_val
+    exp_val = options.exp_val
     round_mode = options.round_mode
     exp_mode = options.exp_mode
-    precision = options.ndigits
-    mantissa, temp_exp, base = get_mantissa_exp_base(num, exp_mode, exp)
-    round_digit = get_round_digit(mantissa, round_mode, precision)
+    ndigits = options.ndigits
+    mantissa, temp_exp_val, base = get_mantissa_exp_base(num, exp_mode,
+                                                         exp_val)
+    round_digit = get_round_digit(mantissa, round_mode, ndigits)
     mantissa_rounded = round(mantissa, -round_digit)
 
     '''
     Repeat mantissa + exponent discovery after rounding in case rounding
     altered the required exponent.
     '''
-    rounded_num = mantissa_rounded * Decimal(base)**Decimal(temp_exp)
-    mantissa, exp, base = get_mantissa_exp_base(rounded_num, exp_mode, exp)
-    round_digit = get_round_digit(mantissa, round_mode, precision)
+    rounded_num = mantissa_rounded * Decimal(base)**Decimal(temp_exp_val)
+    mantissa, exp_val, base = get_mantissa_exp_base(rounded_num, exp_mode,
+                                                    exp_val)
+    round_digit = get_round_digit(mantissa, round_mode, ndigits)
     mantissa_rounded = Decimal(round(mantissa, -int(round_digit)))
 
     if mantissa_rounded == 0:
@@ -103,7 +105,7 @@ def format_num(num: Decimal, unrendered_options: FormatOptions) -> str:
         result is technically correct (e.g. 0e+03 = 0e+00), but sciform
         always presents zero values with an exponent of zero.
         '''
-        exp = 0
+        exp_val = 0
 
     fill_char = options.fill_mode.to_char()
     mantissa_str = format_num_by_top_bottom_dig(mantissa_rounded.normalize(),
@@ -121,7 +123,7 @@ def format_num(num: Decimal, unrendered_options: FormatOptions) -> str:
                                   lower_separator,
                                   group_size=3)
 
-    exp_str = get_exp_str(exp, exp_mode, options.capitalize)
+    exp_str = get_exp_str(exp_val, exp_mode, options.capitalize)
     exp_str = convert_exp_str(exp_str,
                               options.prefix_exp,
                               options.parts_per_exp,
@@ -213,20 +215,20 @@ def format_val_unc(val: Decimal, unc: Decimal,
         exp_driver = unc_rounded
         val_exp_driver = False
 
-    _, exp, _ = get_mantissa_exp_base(
[…]
