+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.25.2"
+__version__ = "0.26.0"
 
 from sciform.format_options import FormatOptions
 from sciform.formatter import Formatter
@@ -20,5 +20,4 @@ __all__ = ['__version__', 'FormatOptions', 'Formatter', 'SciNum', 'SciNumUnc',
            'global_reset_iec_prefixes', 'global_reset_parts_per_forms',
            'print_global_defaults', 'GlobalDefaultsContext', 'FillMode',
            'SignMode', 'GroupingSeparator', 'RoundMode', 'ExpMode',
-           'AutoExpVal',
-           'AutoRound']
+           'AutoExpVal', 'AutoRound']
