### 2023-08-17 0dca634
global config unit tests
