### 2023-08-17 6a1a165
test invalid options
