### 2023-08-17 e016ef2
some binary tests
