### 2023-08-18 17b5275
split tests into separate functions
