### 2023-08-18 42bf6af
break out validate_options and apply to both RenderedFormatOptions and FormatOptions
