### 2023-08-18 f2c8850
remove if __name__ == "__main__" in tests
