+++ tests/test_float_fsml.py
@@ -606,7 +606,3 @@ class TestFormatting(unittest.TestCase):
         ]
 
         self.run_float_fsml_cases(cases_list)
-
-
-if __name__ == "__main__":
-    unittest.main()
