### 2023-08-18 1aff9bf
test_dec_place_auto_round
