### 2023-08-18 af7f116
binary latex
