### 2023-08-18 8999256
test auto round in dec_place mode
