+++ tests/test_val_unc_formatter.py
@@ -1,6 +1,7 @@
 import unittest
 
-from sciform import FormatOptions, Formatter, ExpMode, GroupingSeparator
+from sciform import (FormatOptions, Formatter, ExpMode, GroupingSeparator,
+                     RoundMode)
 
 
 ValUncFormatOptionsCases = list[tuple[tuple[float, float],
@@ -171,3 +172,7 @@ class TestFormatting(unittest.TestCase):
     def test_pdg_ndigits_error(self):
         self.assertRaises(ValueError, FormatOptions, pdg_sig_figs=True,
                           ndigits=0)
+
+    def test_dec_place_warn(self):
+        sform = Formatter(FormatOptions(round_mode=RoundMode.DEC_PLACE))
+        self.assertWarns(Warning, sform, 42, 24)
