### 2023-08-18 2bb4c46
test SciNum and SciNumUnc repr
