+++ src/sciform/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.26.0"
+__version__ = "0.26.1"
 
 from sciform.format_options import FormatOptions
 from sciform.formatter import Formatter
