### 2023-08-18 4350abb
removed expectedFailure from unittest and add docstring explaining the subtle, but expected/desired behavior.
