+++ src/sciform/__init__.py
@@ -11,7 +11,7 @@ from sciform.format_options import (
     print_global_defaults, GlobalDefaultsContext)
 
 from sciform.modes import (FillMode, SignMode, GroupingSeparator, RoundMode,
-                           ExpMode, AutoExpVal, AutoRound)
+                           ExpMode, AutoExpVal, AutoDigits)
 
 __all__ = ['__version__', 'FormatOptions', 'Formatter', 'SciNum', 'SciNumUnc',
            'set_global_defaults', 'reset_global_defaults',
@@ -20,4 +20,4 @@ __all__ = ['__version__', 'FormatOptions', 'Formatter', 'SciNum', 'SciNumUnc',
            'global_reset_iec_prefixes', 'global_reset_parts_per_forms',
            'print_global_defaults', 'GlobalDefaultsContext', 'FillMode',
            'SignMode', 'GroupingSeparator', 'RoundMode', 'ExpMode',
-           'AutoExpVal', 'AutoRound']
+           'AutoExpVal', 'AutoDigits']
