### 2023-08-19 9a780df
examples folder with python scripts, data, figure pngs and table text files
