+++ examples/requirements-example.txt
@@ -1,0 +1,4 @@
+numpy
+scipy
+matplotlib
+tabulate
