+++ examples/outputs/fit_plot_with_sciform_table.txt
@@ -1,0 +1,9 @@
++---------+-----------------+------------------+---------------------+
+| color   | curvature       | x0               | y0                  |
++=========+=================+==================+=====================+
+| red     | (20.7(1.9))e+12 | (-42.7(4.8))e-06 | (1.0000060(46))e+09 |
++---------+-----------------+------------------+---------------------+
+| blue    | (18.4(1.0))e+12 | (-6.9(1.6))e-06  | (1.0000262(29))e+09 |
++---------+-----------------+------------------+---------------------+
+| purple  | (21.7(1.7))e+12 | (15.1(2.5))e-06  | (1.0000246(44))e+09 |
++---------+-----------------+------------------+---------------------+
