### 2023-08-19 ac3400d
Updates examples documentation to use artifacts in examples directory
