### 2023-08-20 6c93547
add repr for FormatOptions and RenderedFormatOptions. Use this for print_global_defaults
