+++ CHANGELOG.rst
@@ -1,7 +1,13 @@
+0.27.2 (2023-08-20)
+-------------------
+
+* Add ``__repr__()`` for ``FormatOptions`` and
+  ``RenderedFormatOptions``.
+
 0.27.1 (2023-08-18)
 -------------------
 
-* Add ``/examples/`` folder to hold example scripts used in the
+* Add ``examples/`` folder to hold example scripts used in the
   documentation as well as the input data for these scripts and their
   outputs which appear in the documentation.
 * Remove extra ``readthedocs.yaml`` file.
