+++ README.rst
@@ -73,12 +73,13 @@ strings according to the selected options.
 For brevity, the user may consider abbreviating ``FormatOptions`` as
 ``Fo``.
 
->>> from sciform import FormatOptions as Fo
->>> from sciform import Formatter, RoundMode, GroupingSeparator, ExpMode
->>> sform = Formatter(Fo(round_mode=RoundMode.DEC_PLACE,
-...                      ndigits=6,
-...                      upper_separator=GroupingSeparator.SPACE,
-...                      lower_separator=GroupingSeparator.SPACE))
+>>> from sciform import (Formatter, FormatOptions, RoundMode,
+>>>                      GroupingSeparator, ExpMode)
+>>> sform = Formatter(FormatOptions(
+...             round_mode=RoundMode.DEC_PLACE,
+...             ndigits=6,
+...             upper_separator=GroupingSeparator.SPACE,
+...             lower_separator=GroupingSeparator.SPACE))
 >>> print(sform(51413.14159265359))
 51 413.141 593
 
@@ -96,13 +97,15 @@ to format pairs of numbers as value/uncertainty pairs.
 This can be done by passing two numbers into a ``Formatter`` call or by
 using the ``SciNumUnc`` object.
 
->>> sform = Formatter(Fo(ndigits=2,
-...                      upper_separator=GroupingSeparator.SPACE,
-...                      lower_separator=GroupingSeparator.SPACE))
+>>> sform = Formatter(FormatOptions(
+...             ndigits=2,
+...             upper_separator=GroupingSeparator.SPACE,
+...             lower_separator=GroupingSeparator.SPACE))
 >>> print(sform(123456.654321, 0.0034))
 123 456.654 3 +/- 0.003 4
->>> sform = Formatter(Fo(ndigits=4,
-...                      exp_mode=ExpMode.ENGINEERING))
+>>> sform = Formatter(FormatOptions(
+...             ndigits=4,
+...             exp_mode=ExpMode.ENGINEERING))
 >>> print(sform(123456.654321, 0.0034))
 (123.456654321 +/- 0.000003400)e+03
 
