### 2023-08-23 df02472
remove __version__, set in pyproject.toml, use importlib.metadata in conf.py
