+++ pyproject.toml
@@ -1,10 +1,9 @@
 [build-system]
-requires = ["setuptools>=61.0"]
+requires = ["setuptools>=61.0", "setuptools_scm[toml]>=6.2"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "sciform"
-version = "0.27.3"
 authors = [
     {name = "Justin Gerber", email="justin.gerber48@gmail.com"}
 ]
@@ -12,7 +11,7 @@ description = "A package for formatting numbers into scientific formatted string
 classifiers = [
     "License :: OSI Approved :: MIT License"
 ]
-dynamic = ["readme"]
+dynamic = ["version", "readme"]
 
 [project.urls]
 homepage = "https://github.com/jagerber48/sciform"
@@ -20,3 +19,5 @@ documentation = "https://sciform.readthedocs.io"
 
 [tool.setuptools.dynamic]
 readme = {file = "README.rst"}
+
+[tool.setuptools_scm]
