+++ docs/source/exp replacement.rst
@@ -149,8 +149,8 @@ Note that it is possible, using the ``extra_parts_per_forms`` option to
 override the standard mappings listed above.
 
 >>> from sciform import Formatter, FormatOptions, ExpMode
->>> sform = Formatter(Fo(
-                exp_mode=ExpMode.ENGINEERING,
+>>> sform = Formatter(FormatOptions(
+...             exp_mode=ExpMode.ENGINEERING,
 ...             parts_per_exp=True,
 ...             extra_parts_per_forms={-9: None, -12: 'ppb'}))
 >>> print(sform(33e-9))
