+++ .github/workflows/python-publish.yml
@@ -35,5 +35,5 @@ jobs:
     - name: Publish package
       uses: pypa/gh-action-pypi-publish@27b31702a0e7fc50959f5ad993c78deac1bdfc29
       with:
-        user: __token__
+        user: Jagerber48
         password: ${{ secrets.PYPI_API_TOKEN }}
