+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Unreleased
 ----------
 
+0.27.5 (2023-08-25)
+-------------------
+
+* Fix auto-publish.
+
 0.27.4 (2023-08-25)
 -------------------
 
