### 2023-08-26 6781e1a
formatutils, clean up some typically unreachable code for testing
