+++ tests/test_invalid_options.py
@@ -1,7 +1,14 @@
 import unittest
+from decimal import Decimal
 
 from sciform import (Formatter, FormatOptions, RoundMode, ExpMode,
                      GroupingSeparator)
+from sciform.formatting import format_non_inf
+from sciform.grouping import add_group_chars_between_numbers, add_separators
+from sciform.format_utils import (
+    get_top_digit, get_mantissa_exp_base, get_exp_str, get_sign_str,
+    get_round_digit
+)
 
 
 class TestInvalidOptions(unittest.TestCase):
@@ -127,3 +134,64 @@ class TestInvalidOptions(unittest.TestCase):
             prefix_exp=True,
             parts_per_exp=True
         )
+
+    def test_format_non_inf(self):
+        self.assertRaises(ValueError, format_non_inf, Decimal(1.0),
+                          FormatOptions().render())
+
+    def test_add_group_chars(self):
+        self.assertRaises(ValueError, add_group_chars_between_numbers,
+                          string='123456.654321', group_char='_',
+                          direction='forwards', group_size=3)
+
+    def test_add_separators(self):
+        self.assertRaises(ValueError, add_separators,
+                          num_str='123.456.789',
+                          upper_separator=',',
+                          decimal_separator='.',
+                          lower_separator='_',
+                          group_size=3)
+
+    def test_get_top_digit_infinite(self):
+        self.assertEqual(get_top_digit(Decimal('nan')), 0)
+
+    def test_get_mantissa_exp_base_fixed_point_set_exp(self):
+        self.assertRaises(ValueError, get_mantissa_exp_base,
+                          num=Decimal(3),
+                          exp_mode=ExpMode.FIXEDPOINT,
+                          input_exp_val=1)
+
+    def test_get_mantissa_exp_base_engineering_set_exp(self):
+        self.assertRaises(ValueError, get_mantissa_exp_base,
+                          num=Decimal(3),
+                          exp_mode=ExpMode.ENGINEERING,
+                          input_exp_val=1)
+
+    def test_get_mantissa_exp_base_binary_iec_set_exp(self):
+        self.assertRaises(ValueError, get_mantissa_exp_base,
+                          num=Decimal(3),
+                          exp_mode=ExpMode.BINARY_IEC,
+                          input_exp_val=3)
+
+    def test_get_mantissa_exp_base_bad_exp_mode(self):
+        self.assertRaises(ValueError, get_mantissa_exp_base,
+                          num=Decimal(3),
+                          exp_mode='eng',
+                          input_exp_val=3)
+
+    def test_get_exp_str_bad_exp_mode(self):
+        self.assertRaises(ValueError, get_exp_str,
+                          exp_val=2,
+                          exp_mode='sci',
+                          capitalize=False)
+
+    def test_get_sign_str_bad_sign_mode(self):
+        self.assertRaises(ValueError, get_sign_str,
+                          num=Decimal(1),
+                          sign_mode='space')
+
+    def test_get_round_digit_bad_round_mode(self):
+        self.assertRaises(ValueError, get_round_digit,
+                          num=Decimal(123.456),
+                          round_mode='none',
+                          ndigits=0)
