### 2023-08-26 d0ef4c1
Refactor how exp string is resolved
changes behavior on nan formatting with % mode
also use \text{nan} for latex with nan or inf
