+++ src/sciform/formatting.py
@@ -2,18 +2,16 @@ from warnings import warn
 import re
 from decimal import Decimal
 
-from sciform.modes import ExpMode, SignMode, AutoExpVal, RoundMode
+from sciform.modes import ExpMode, SignMode, AutoExpVal, RoundMode, ExpFormat
 from sciform.format_options import FormatOptions, RenderedFormatOptions
-from sciform.format_utils import (get_mantissa_exp_base, get_exp_str,
-                                  get_top_digit,
-                                  get_round_digit,
-                                  format_num_by_top_bottom_dig,
-                                  convert_exp_str,
-                                  latex_translate)
+from sciform.format_utils import (
+    get_mantissa_exp_base, get_exp_str, get_top_digit, get_round_digit,
+    format_num_by_top_bottom_dig, latex_translate, parse_standard_exp_str
+)
 from sciform.grouping import add_separators
 
 
-def format_non_inf(num: Decimal, options: RenderedFormatOptions) -> str:
+def format_non_finite(num: Decimal, options: RenderedFormatOptions) -> str:
     if num.is_nan():
         num_str = 'nan'
     elif num == Decimal('inf'):
@@ -21,7 +19,7 @@ def format_non_inf(num: Decimal, options: RenderedFormatOptions) -> str:
     elif num == Decimal('-inf'):
         num_str = '-inf'
     else:
-        raise ValueError(f'format_non_inf() cannot format {num}.')
+        raise ValueError(f'format_non_finite() cannot format {num}.')
 
     if options.nan_inf_exp:
         exp_mode = options.exp_mode
@@ -30,23 +28,15 @@ def format_non_inf(num: Decimal, options: RenderedFormatOptions) -> str:
         if options.exp_val is AutoExpVal:
             exp_val = 0
 
-        if exp_mode is ExpMode.FIXEDPOINT:
-            exp_str = ''
-        elif (exp_mode is ExpMode.SCIENTIFIC
-              or exp_mode is ExpMode.ENGINEERING
-              or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-            exp_str = f'e+{exp_val:02d}'
-        else:
-            exp_str = f'b+{exp_val:02d}'
-
-        exp_str = convert_exp_str(exp_str,
-                                  options.prefix_exp,
-                                  options.parts_per_exp,
-                                  options.latex,
-                                  options.superscript_exp,
-                                  options.extra_si_prefixes,
-                                  options.extra_iec_prefixes,
-                                  options.extra_parts_per_forms)
+        exp_str = get_exp_str(exp_val,
+                              exp_mode,
+                              options.exp_format,
+                              options.capitalize,
+                              options.latex,
+                              options.superscript_exp,
+                              options.extra_si_prefixes,
+                              options.extra_iec_prefixes,
+                              options.extra_parts_per_forms)
     else:
         exp_str = ''
 
@@ -54,9 +44,6 @@ def format_non_inf(num: Decimal, options: RenderedFormatOptions) -> str:
         result = f'({num_str}){exp_str}'
     else:
         result = f'{num_str}'
-
-    if options.exp_mode is ExpMode.PERCENT:
-        result = f'({result})%'
 
     if options.capitalize:
         result = result.upper()
@@ -73,7 +60,7 @@ def format_num(num: Decimal, unrendered_options: FormatOptions) -> str:
     options = unrendered_options.render()
 
     if not num.is_finite():
-        return format_non_inf(num, options)
+        return format_non_finite(num, options)
 
     if options.exp_mode is ExpMode.PERCENT:
         num *= 100
@@ -123,20 +110,17 @@ def format_num(num: Decimal, unrendered_options: FormatOptions) -> str:
[…]
