+++ src/sciform/format_utils.py
@@ -1,10 +1,10 @@
 from typing import Union
 from math import floor, log10, log2
 import re
-from copy import copy
 from decimal import Decimal
 
-from sciform.modes import ExpMode, RoundMode, SignMode, AutoExpVal, AutoDigits
+from sciform.modes import (ExpMode, ExpFormat, RoundMode, SignMode, AutoExpVal,
+                           AutoDigits)
 from sciform.prefix import (si_val_to_prefix_dict, iec_val_to_prefix_dict,
                             pp_val_to_prefix_dict)
 
@@ -99,22 +99,127 @@ def get_mantissa_exp_base(
     return mantissa, exp_val, base
 
 
-def get_exp_str(exp_val: int, exp_mode: ExpMode,
-                capitalize: bool) -> str:
-    if exp_mode is exp_mode.FIXEDPOINT or exp_mode is ExpMode.PERCENT:
-        exp_str = ''
-    elif (exp_mode is ExpMode.SCIENTIFIC
-          or exp_mode is ExpMode.ENGINEERING
-          or exp_mode is ExpMode.ENGINEERING_SHIFTED):
-        exp_char = 'E' if capitalize else 'e'
-        exp_str = f'{exp_char}{exp_val:+03d}'
+def get_standard_exp_str(base: int,
+                         exp_val: int,
+                         capitalize: bool) -> str:
+    base_exp_symb_dict = {10: 'e', 2: 'b'}
+    exp_symb = base_exp_symb_dict[base]
+    if capitalize:
+        exp_symb = exp_symb.capitalize()
+    return f'{exp_symb}{exp_val:+03d}'
+
+
+def get_superscript_exp_str(base: int,
+                            exp_val: int) -> str:
+    sup_trans = str.maketrans("+-0123456789", "⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹")
+    exp_val_str = f'{exp_val}'.translate(sup_trans)
+    return f'×{base}{exp_val_str}'
+
+
+def get_prefix_dict(exp_format: ExpFormat,
+                    base: int,
+                    extra_si_prefixes: dict[int, str] = None,
+                    extra_iec_prefixes: dict[int, str] = None,
+                    extra_parts_per_forms: dict[int, str] = None
+                    ) -> dict[int, str]:
+    if exp_format is ExpFormat.PREFIX:
+        if base == 10:
+            prefix_dict = si_val_to_prefix_dict.copy()
+            prefix_dict.update(extra_si_prefixes)
+        elif base == 2:
+            prefix_dict = iec_val_to_prefix_dict.copy()
+            prefix_dict.update(extra_iec_prefixes)
+        else:
+            raise ValueError(f'Unhandled base {base}')
+    elif exp_format is ExpFormat.PARTS_PER:
+        prefix_dict = pp_val_to_prefix_dict.copy()
+        prefix_dict.update(extra_parts_per_forms)
+    else:
+        raise ValueError(f'Unhandled ExpFormat, {exp_format}.')
+
+    return prefix_dict
+
+
+def get_exp_str(exp_val: int,
+                exp_mode: ExpMode,
+                exp_format: ExpFormat,
+                capitalize: bool,
+                latex: bool,
+                superscript: bool,
+                extra_si_prefixes: dict[int, str] = None,
+                extra_iec_prefixes: dict[int, str] = None,
+                extra_parts_per_forms: dict[int, str] = None) -> str:
+    if exp_mode is ExpMode.FIXEDPOINT:
+        return ''
+    elif exp_mode is ExpMode.PERCENT:
+        return '%'
+
+    if (exp_mode is ExpMode.SCIENTIFIC
[…]
