+++ tests/test_print.py
@@ -21,8 +21,8 @@ class TestPrint(unittest.TestCase):
             " 'lower_separator': <GroupingSeparator.NONE: 'no_grouping'>,\n"
             " 'sign_mode': <SignMode.NEGATIVE: 'negative'>,\n"
             " 'fill_mode': <FillMode.SPACE: 'space'>,\n"
-            " 'top_dig_place': 0,\n 'prefix_exp': False,\n"
-            " 'parts_per_exp': False,\n"
+            " 'top_dig_place': 0,\n"
+            " 'exp_format': <ExpFormat.STANDARD: 'standard'>,\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
@@ -54,8 +54,8 @@ class TestPrint(unittest.TestCase):
             " 'lower_separator': <GroupingSeparator.NONE: 'no_grouping'>,\n"
             " 'sign_mode': <SignMode.NEGATIVE: 'negative'>,\n"
             " 'fill_mode': <FillMode.SPACE: 'space'>,\n"
-            " 'top_dig_place': 0,\n 'prefix_exp': False,\n"
-            " 'parts_per_exp': False,\n"
+            " 'top_dig_place': 0,\n"
+            " 'exp_format': <ExpFormat.STANDARD: 'standard'>,\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"

[…]
