+++ docs/source/usage.rst
@@ -153,8 +153,7 @@ package default settings):
  'sign_mode': <SignMode.NEGATIVE: 'negative'>,
  'fill_mode': <FillMode.SPACE: 'space'>,
  'top_dig_place': 0,
- 'prefix_exp': False,
- 'parts_per_exp': False,
+ 'exp_format': <ExpFormat.STANDARD: 'standard'>,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'extra_parts_per_forms': {},
@@ -193,8 +192,7 @@ unchanged.
  'sign_mode': <SignMode.NEGATIVE: 'negative'>,
  'fill_mode': <FillMode.ZERO: 'zero'>,
  'top_dig_place': 0,
- 'prefix_exp': False,
- 'parts_per_exp': False,
+ 'exp_format': <ExpFormat.STANDARD: 'standard'>,
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'extra_parts_per_forms': {},
