### 2023-08-26 546c035
latex trim whitespace option
