### 2023-08-27 73d4813
catch superscript exponent while in prefix mode when there is no replacement
