### 2023-08-27 b04121f
no cover for unreachable code
