### 2023-08-27 356ff17
tests for unsupported binary value/uncertainty formatting
