### 2023-08-27 24eea38
test for reformatted exponent string resolution
