+++ src/sciform/formatting.py
@@ -139,8 +139,8 @@ def format_val_unc(val: Decimal, unc: Decimal,
     options = unrendered_options.render()
     if (options.exp_mode is ExpMode.BINARY
             or options.exp_mode is ExpMode.BINARY_IEC):
-        raise NotImplementedError(f'Binary exponent modes are not supported '
-                                  f'for value/uncertainty formatting.')
+        raise NotImplementedError('Binary exponent modes are not supported '
+                                  'for value/uncertainty formatting.')
 
     if options.round_mode is RoundMode.DEC_PLACE:
         warn('Precision round mode not available for value/uncertainty '
