### 2023-08-27 ec2dd9d
Optional type hints for all FormatOptions inputs
