### 2023-08-27 3734243
Update fsml documentation to point to existing format options documentation to avoid duplication
