+++ docs/source/fsml.rst
@@ -32,9 +32,16 @@ The :mod:`sciform` format specification mini-language is given by::
     exp_mode           ::=  "f" | "F" | "%" | "e" | "E" | "r" | "R" | "b" | "B" |
     exp_val            ::=  [+-]?digit+
 
+Example:
+
+>>> from sciform import SciNum
+>>> print(f'{SciNum(123456.654321):_,_.4}')
+123_456,654_3
+
 See below for details about how the terms in the FSML correspond to
 formatting options.
-For further details about the options see :ref:`formatting_options`.
+Further details about the options can be found at
+:ref:`formatting_options`.
 
 .. list-table:: :mod:`sciform` Format Specification Mini-Language Terms
    :widths: 15 30
@@ -44,114 +51,79 @@ For further details about the options see :ref:`formatting_options`.
      - Description
    * - | fill
        | (``'0='``, ``' ='``)
-     - Fill characters will be padded between the most signifant digit
-       and the sign symbol until the digit corresponding to the
-       ``fill_top_digit`` is filled.
+     - Configure ``fill_mode`` to be :class:`FillMode.ZERO` or
+       :class:`FillMode.SPACE`. See :ref:`left_filling`.
    * - | sign
        | (``'-'``, ``'+'``, ``' '``)
-     - ``'-'`` will include a sign symbol only for negative numbers.
-       ``'+'`` will include a sign symbol for all numbers.
-       ``' '`` will include a minus symbol for negative numbers and a
-       space for positive numbers. Zero is always considered to be
-       positive.
+     - Configure ``sign_mode`` to be :class:`SignMode.NEGATIVE`,
+       :class:`SignMode.ALWAYS` or :class:`SignMode.SPACE`. See
+       :ref:`sign_mode`.
    * - | alternate mode
        | (``'#'``)
-     - Alternate mode is enabled (disabled by default) if the ``'#'``
-       flag is included in the format specification.
-       In engineering mode (``r`` or ``R``), the alternate mode engages
-       :ref:`engineering_shifted` mode.
-       In binary mode (``b`` or ``B``), the alternate mode engages
-       :ref:`binary_iec` mode.
+     - The alternate mode flag indicates to use
+       :ref:`engineering_shifted` mode when the exponent mode flag is
+       ``'r'`` or ``'R'`` or to use :ref:`binary_iec` mode when the
+       exponent mode flag is ``'b'`` or ``'B'``.
    * - | fill_top_digit
        | (``\d+``)
-     - Any non-negative integer, default (0).
-       Indicates the decimal or binary place to which the formatted
-       string should be padded.
-       e.g. ``f'{SciNum(123):0=4}'`` will give ``'00123'``, i.e. padding
-       to the ``10^4`` place.
+     - Sets ``top_dig_place`` to any non-negative integer.
+       See :ref:`left_filling`.
    * - | upper_separator
        | (``'n'``, ``','``, ``'.'``, ``'s'``, ``'_'``)
-     - Indicates the character to use as the separator between groups of
-       three digits above the decimal point. For base 10 formats this is
-       the "thousands" separator.
-       ``'n'`` is no separator, ``'s'`` is a single-whitespace separator
-       and ``','``, ``'.'``, and ``'_'`` are comma, period, and
-       underscore separators.
-       Note that NIST discourages the use of ``','`` or ``'.'`` as
-       thousands seperators because they can be confused with the
-       decimal separators depending on the locality.
-       See
[…]
