+++ .github/workflows/python-package.yml
@@ -38,3 +38,7 @@ jobs:
     - name: Test with pytest
       run: |
         pytest --doctest-glob="*.rst"
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v3
+      env:
+        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
