+++ .github/workflows/python-package.yml
@@ -37,7 +37,8 @@ jobs:
         flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
     - name: Test with pytest
       run: |
-        pytest --doctest-glob="*.rst" --cov
+        pytest --doctest-glob="*.rst"
+        pytest --cov .
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
       env:
