+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Unreleased
   pass type checks.
 * Write format-specification mini-language documentation to refer to
   existing format options to avoid documentation duplication.
+* Perform test coverage report and upload to ``codecov.io``.
 
 0.28.0 (2023-08-27)
 -------------------
