+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Unreleased
 * Write format-specification mini-language documentation to refer to
   existing format options to avoid documentation duplication.
 * Perform test coverage report and upload to `<https://codecov.io>`_.
-  Add badge to readme.
+* Add package status badges to readme.
 
 0.28.0 (2023-08-27)
 -------------------
