### 2023-08-28 922f36c
list version in pyproject.toml
