### 2023-08-29 b1dd622
usage.rst cleanup
