### 2023-08-31 f30bbeb
index on bugfix/pdg_decimal_rounding: a1a14d9 decimal math
