+++ CHANGELOG.rst
@@ -1,14 +1,10 @@
 Unreleased
-==========
+----------
 
-Fixed
------
-
-* Fixed a bug when using ``pdg_sig_figs`` with 
-
+* No unreleased changes.
 
 0.28.1 (2023-08-28)
-===================
+-------------------
 
 * Make ``FormatOptions`` inputs ``Optional`` so that ``None`` inputs
   pass type checks.
@@ -22,7 +18,7 @@ Fixed
 * List supported Python versions in ``pyproject.toml`` classifiers.
 
 0.28.0 (2023-08-27)
-===================
+-------------------
 
 * **[BREAKING]** Replace ``prefix_exp`` and ``parts_per_exp`` options
   with an ``exp_format`` option which can be configured to
@@ -48,12 +44,12 @@ Fixed
   yet.
 
 0.27.4 (2023-08-25)
-===================
+-------------------
 
 * Setup github action to automatically build and publish on release.
 
 0.27.3 (2023-08-23)
-===================
+-------------------
 
 * Added ``Unreleased`` section to changelog.
 * Removed ``version`` from source code.
@@ -62,13 +58,13 @@ Fixed
 * Stopped encouraging ``import FormatOptions as Fo``.
 
 0.27.2 (2023-08-20)
-===================
+-------------------
 
 * Add ``__repr__()`` for ``FormatOptions`` and
   ``RenderedFormatOptions``.
 
 0.27.1 (2023-08-18)
-===================
+-------------------
 
 * Add ``examples/`` folder to hold example scripts used in the
   documentation as well as the input data for these scripts and their
@@ -76,7 +72,7 @@ Fixed
 * Remove extra ``readthedocs.yaml`` file.
 
 0.27.0 (2023-08-18)
-===================
+-------------------
 
 * **[BREAKING]** Rename ``AutoRound`` to ``AutoDigits``. This is
   because, e.g., ``ndigits=AutoDigits`` sounds more correct than
@@ -84,7 +80,7 @@ Fixed
   confused as being an option for ``round_mode``, which it is not.
 
 0.26.2 (2023-08-18)
-===================
+-------------------
 
 * Fix a bug where illegal options combinations could be realized at
   format time when certain global default objects were merged into
@@ -93,12 +89,12 @@ Fixed
   in the global defaults but before formatting.
 
 0.26.1 (2023-08-18)
-===================
+-------------------
 
 * Add unit tests, increase test coverage.
 
 0.26.0 (2023-08-15)
-===================
+-------------------
 
 * **[BREAKING]** Rename some format options to make their usage more
   clear.
@@ -125,17 +121,17 @@ Fixed
    * More accurate descriptions of some invalid options combinations.
 
 0.25.2 (2023-08-11)
-===================
+-------------------
 
 * Update roadmap
 
 0.25.1 (2023-08-10)
-===================
+-------------------
 
 * Refactor ``get_pdg_round_digit()`` into a dedicated function.
 
 0.25.0 (2023-08-02)
-===================
+-------------------
 
 * **[BREAKING]** ``template`` option removed from ``FormatOptions``
   constructor.
@@ -144,14 +140,14 @@ Fixed
 * Minor documentation improvements.
 
 0.24.0 (2023-07-30)
-===================
+-------------------
 
 * **[BREAKING]** percent mode is now accessed via an exponent mode,
   ``ExpMode.PERCENT``.
   There is no longer a ``percent`` keyword argument.
 
 0.23.0 (2023-07-29)
-===================
+-------------------
 
 * **[BREAKING]** Users now construct ``FormatOptions`` objects which
   they pass into ``Formatter`` objects and global configuration
@@ -166,18 +162,18 @@ Fixed
 * Change ``pyproject.toml`` description
 
 0.22.2 (2023-07-27)
[…]
