### 2023-08-31 f8784b9
exp_format defaults to None
