### 2023-09-02 0b7c890
No parentheses around bracket uncertainty strings with alphabetic exponent replacements.
