### 2023-09-02 c047fe5
test no prefix for exp_val=0
