### 2023-09-04 d0a93a7
Change strategy for when to exclude wrapping parentheses
