+++ tests/test_val_unc_formatter.py
@@ -69,6 +69,10 @@ class TestFormatting(unittest.TestCase):
                                exp_format=ExpFormat.PARTS_PER),
                  '123.456(789) ppm'),
                 (FormatOptions(bracket_unc=True,
+                               exp_mode=ExpMode.ENGINEERING,
+                               exp_format=ExpFormat.PREFIX),
+                 '123.456(789) μ'),
+                (FormatOptions(bracket_unc=True,
                                exp_mode=ExpMode.ENGINEERING_SHIFTED,
                                exp_format=ExpFormat.PARTS_PER,
                                add_ppth_form=True),
