+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+This project adheres to `Semantic Versioning <https://semver.org/>`_.
+
+----
+
 Unreleased
 ----------
 
@@ -23,6 +27,8 @@ Improved
 ^^^^^^^^
 
 * Fixed minor typos in documentation.
+
+----
 
 0.28.2 (2023-08-31)
 -------------------
@@ -51,6 +57,8 @@ Fixed
   In the future a flag may be added to select "parts-per" formatting
   using the format specification mini-language.
 
+----
+
 0.28.1 (2023-08-28)
 -------------------
 
@@ -64,6 +72,8 @@ Fixed
 * Add package status badges to readme.
 * Test against Python 3.11.
 * List supported Python versions in ``pyproject.toml`` classifiers.
+
+----
 
 0.28.0 (2023-08-27)
 -------------------
@@ -91,10 +101,14 @@ Fixed
   Rounding and truncating are not properly implemented in binary mode
   yet.
 
+----
+
 0.27.4 (2023-08-25)
 -------------------
 
 * Setup github action to automatically build and publish on release.
+
+----
 
 0.27.3 (2023-08-23)
 -------------------
@@ -105,11 +119,15 @@ Fixed
   ``setuptools_scm``.
 * Stopped encouraging ``import FormatOptions as Fo``.
 
+----
+
 0.27.2 (2023-08-20)
 -------------------
 
 * Add ``__repr__()`` for ``FormatOptions`` and
   ``RenderedFormatOptions``.
+
+----
 
 0.27.1 (2023-08-18)
 -------------------
@@ -119,6 +137,8 @@ Fixed
   outputs which appear in the documentation.
 * Remove extra ``readthedocs.yaml`` file.
 
+----
+
 0.27.0 (2023-08-18)
 -------------------
 
@@ -126,6 +146,8 @@ Fixed
   because, e.g., ``ndigits=AutoDigits`` sounds more correct than
   ``ndigits=AutoRound``. Furthermore, ``AutoRound`` could likely be
   confused as being an option for ``round_mode``, which it is not.
+
+----
 
 0.26.2 (2023-08-18)
 -------------------
@@ -136,10 +158,14 @@ Fixed
   The bug is fixed by re-checking the options combinations after merging
   in the global defaults but before formatting.
 
+----
+
 0.26.1 (2023-08-18)
 -------------------
 
 * Add unit tests, increase test coverage.
+
+----
 
 0.26.0 (2023-08-15)
 -------------------
@@ -168,15 +194,21 @@ Fixed
    * Better explanations of ``AutoExpVal`` and ``AutoRound`` behavior.
    * More accurate descriptions of some invalid options combinations.
 
+----
+
 0.25.2 (2023-08-11)
 -------------------
 
 * Update roadmap
 
+----
+
 0.25.1 (2023-08-10)
 -------------------
 
 * Refactor ``get_pdg_round_digit()`` into a dedicated function.
+
+----
 
 0.25.0 (2023-08-02)
 -------------------
@@ -187,12 +219,16 @@ Fixed
   ``FormatOptions`` instances using the ``merge()`` method.
 * Minor documentation improvements.
 
+----
+
 0.24.0 (2023-07-30)
 -------------------
 
 * **[BREAKING]** percent mode is now accessed via an exponent mode,
   ``ExpMode.PERCENT``.
   There is no longer a ``percent`` keyword argument.
+
+----
 
 0.23.0 (2023-07-29)
 -------------------
@@ -209,16 +245,22 @@ Fixed
   ``SciNumUnc`` objects.
 * Change ``pyproject.toml`` description
 
+----
+
 0.22.2 (2023-07-27)
 -------------------
 
[…]
